Module star_tracker_lib::attitude_determination

attitude_determination is the process of voting for the most likely attitude, this is the final step in the process.

The following operations would have already been executed:

reading image.

thresholding.

finding blobs.

finding unique star pairs.

This stage uses given sensor position and given actual position of any positioning system to estimate the most likely attitude.
A rotation matrix/vector is generated by the determination system which can be applied to any local point and it will be converted to the world position.

Triad

+ Easy to implement.
- Only works with 2 vectors.
- Only considers 1.5 sensor values.

This method only allows 1.5 measurements and is not ideal for a star tracker.
The methodology is to get 2, 3D vector measurements in its local and global frame.
The most accurate vector is assigned to an axis of the rotation matrix.
The least accurate vector is then crossed with the most accurate vector to give another axis.
The last axis is generated by crossing the previous 2 axis.
As this only accounts for < 2 input vectors, it is not ideal for a star tracker.

Davenports Q Method

+ Can have infinite elements.
+ Does not have singularities (always works).
- Very slow (calculates eigenvalues).
- Old, was used a lot but not used now.
- Confusing and lack of instructions.

This method is far more accurate then the triad method and allows infinite elements and considers every element with a given weighting.
It is an old method which uses a set of complicated equations to calculate Wahba’s problem. These equations end up requiring the calculation of an eigenvalue/eigenvector pair from a 4x4 matrix which makes this method undesirable over others.

QUEST

+ Most common method today.
+ A lot faster than davenports method but similar logic.
- Potential singularity (CRP coordinates). <- Can be solved.
- Confusing and lack of instructions.

This method uses iteration to find the eigenvalue/eigenvectors for davenports method.
Unfortuantly at some point, the coordinate system is converted to CRP which has a singularity.

ESOQ

+ No singularity
+ A lot faster than davenport but same logic.
! Made by Mortari (he has done a lot of research into attitude determination, however it is hard to understand his reports).
- Confusing and lack of instructions.

OLEA

How To Use

use star_tracker_lib::attitude_determination::AttitudeDetermination;
use star_tracker_lib::attitude_determination::Quest;
use star_tracker_lib::util::units::Quaternion;
use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Match;

// If you setup projection correctly, the points will be relative to the front of the spacecraft.
// In this situation, I decided that was z: +1.
let reference_forward = Vector3{x: 0.0, y: 0.0, z: 1.0};

// Lets just assume these points are from a valid constellation from tracking_mode.
let found_stars = vec!
[
    // The weight is relative to the other points, the higher the weight the more reliable.
	Match{input: Vector3{x: 0.0, y: 0.0, z: 1.0}, output: Vector3{x: 0.0, y: 0.0, z: 1.0}, weight: 1.0},
	Match{input: Vector3{x: 0.0, y: 1.0, z: 0.0}, output: Vector3{x: 0.0, y: 1.0, z: 0.0}, weight: 1.0},
	Match{input: Vector3{x: 1.0, y: 0.0, z: 0.0}, output: Vector3{x: 1.0, y: 0.0, z: 0.0}, weight: 1.0},
	Match{input: Vector3{x: 0.0, y: 1.0, z: 1.0}, output: Vector3{x: 0.0, y: 1.0, z: 1.0}, weight: 1.0},
];

let rotate_to_cam  : Quaternion = Quest::estimate(&found_stars, None); // Quaternion to rotate world space to camera space.
let rotate_to_world: Quaternion = rotate_to_cam.conjugate();           // Quaternion to rotate camera space to world space.
let world_center = rotate_to_world.rotate_point(reference_forward);    // Rotate spacecraft front to world coordinates

Modules

quest

This contains the estimate method, use this to obtain the attitude of the camera.

wahba

The following equations are essential to the implementation of any algorithms using Wahba’s method.

Structs

Quest

Implements AttitudeDetermination through the quest method.
Use estimate to estimate the attitude of the camera.

Wahba

Formula required for any Wahba based methods.
Does not solve the problem.

Traits

AttitudeDetermination