Module star_tracker_lib::core_include

core_include is All the no_std packages.

This module is found in lib.rs.
Since I am using no_std, some things don’t carry over natively so this module automatically includes them.

Re-exports

pub use crate::util::Maths;

pub use core::result::Result::Ok;

pub use core::result::Result::Err;

pub use core::option::Option::Some;

pub use core::option::Option::None;

Structs

Range

A (half-open) range bounded inclusively below and exclusively above (start..end).

RangeInclusive

A range bounded inclusively below and above (start..=end).

Enums

Option

The Option type. See the module level documentation for more.

Result

Result is a type that represents either success (Ok) or failure (Err).

Traits

Clone

A common trait for the ability to explicitly duplicate an object.

Copy

Types whose values can be duplicated simply by copying bits.

Debug

? formatting.

Eq

Trait for equality comparisons which are equivalence relations.

PartialEq

Trait for equality comparisons.

PartialOrd

Trait for types that form a partial order.

Sized

Types with a constant size known at compile time.

Attribute Macros

derive

Attribute macro used to apply derive macros.

Derive Macros

Clone

Derive macro generating an impl of the trait Clone.

Copy

Derive macro generating an impl of the trait Copy.

Debug

Derive macro generating an impl of the trait Debug.

Eq

Derive macro generating an impl of the trait Eq.

PartialEq

Derive macro generating an impl of the trait PartialEq. The behavior of this macro is described in detail here.

PartialOrd

Derive macro generating an impl of the trait PartialOrd. The behavior of this macro is described in detail here.