Trait star_tracker_lib::image_processing::image::Image

pub trait Image {
    // Required methods
    fn get(&self, px: Pixel) -> Byte;
    fn set(&mut self, px: Pixel, value: Byte);
    fn width(&self) -> usize;
    fn height(&self) -> usize;

    // Provided methods
    fn reset(&mut self) { ... }
    fn valid_pixel(&self, px: Pixel) -> bool { ... }
    fn histogram(&self, histogram: &mut [UInt]) -> Error<()> { ... }
    fn copy_to(&self, to: &mut dyn Image) -> Error<()> { ... }
}

Image accessors, investigation and modification.

Allows for external image packages to be wrapped and new image methods to be investigated. (This is ultimately just to stop requiring generic arguments as consts in generic is too new).

Required Methods

fn get(&self, px: Pixel) -> Byte

Returns the pixel value at the current position.

Arguments

px - The pixel to modify.

fn set(&mut self, px: Pixel, value: Byte)

Sets the pixel value at the current position.

Arguments

px - The pixel to modify.

value - The value to set.

fn width(&self) -> usize

Returns the width of the image.

Returns

The width of the image.

fn height(&self) -> usize

Returns the height of the image.

Returns

The height of the image.

Provided Methods

fn reset(&mut self)

Sets all the pixels to 0. (Allows you to reuse the same image).

Example

use star_tracker_lib::util::units::Pixel;
use star_tracker_lib::image_processing::BasicImage;
use star_tracker_lib::image_processing::Image;

let mut img : BasicImage<3, 3> = BasicImage::new();
img.set(Pixel{x: 0, y: 0}, 10);
img.set(Pixel{x: 2, y: 0}, 20);
img.set(Pixel{x: 0, y: 2}, 32);
img.set(Pixel{x: 0, y: 1}, 43);

img.reset(); // All the pixels are now 0.

fn valid_pixel(&self, px: Pixel) -> bool

True if the get/set will not cause a panic.

Arguments

px - The pixel to modify.

Returns

True if access is safe.

fn histogram(&self, histogram: &mut [UInt]) -> Error<()>

Generates a brightness histogram of the image.

Arguments

histogram - The histogram to fill, must have a size > 0 && < 256.

Example

use star_tracker_lib::image_processing::BasicImage;
use star_tracker_lib::image_processing::Image;
use star_tracker_lib::util::aliases::UInt;
use star_tracker_lib::util::aliases::Byte;
use star_tracker_lib::util::units::Pixel;
let mut img : BasicImage<3, 3> = BasicImage::new();
let mut hist : [UInt; Byte::max_value() as usize + 1] = [0; Byte::MAX as usize + 1];
img.set(Pixel{x: 0, y: 0}, 10);
img.set(Pixel{x: 2, y: 0}, 20);
img.set(Pixel{x: 0, y: 2}, 32);
img.set(Pixel{x: 0, y: 1}, 43);

img.histogram(&mut hist);
assert_eq!(hist[0], 5);
assert_eq!(hist[10], 1);
assert_eq!(hist[20], 1);
assert_eq!(hist[32], 1);
assert_eq!(hist[43], 1);

fn copy_to(&self, to: &mut dyn Image) -> Error<()>

Copies an image over to another image. The is mainly for nix. This can be useful to duplicate an image before blob detection.

Arguments

to - The image to be overwritten.

Returns

Error if invalid.

Example

use star_tracker_lib::util::units::Pixel;
use star_tracker_lib::image_processing::BasicImage;
use star_tracker_lib::image_processing::Image;

let mut from: BasicImage<2, 2> = BasicImage::new();
from.set(Pixel{x: 0, y: 0}, 1);
from.set(Pixel{x: 0, y: 1}, 2);
from.set(Pixel{x: 1, y: 0}, 3);
from.set(Pixel{x: 1, y: 1}, 4);
let mut to: BasicImage<2, 2> = BasicImage::new();
assert!(from.copy_to(&mut to).is_ok());
assert_eq!(to.get(Pixel{x: 0, y: 0}), from.get(Pixel{x: 0, y: 0}));
assert_eq!(to.get(Pixel{x: 0, y: 1}), from.get(Pixel{x: 0, y: 1}));
assert_eq!(to.get(Pixel{x: 1, y: 0}), from.get(Pixel{x: 1, y: 0}));
assert_eq!(to.get(Pixel{x: 1, y: 1}), from.get(Pixel{x: 1, y: 1}));