Struct star_tracker_lib::image_processing::BasicImage

pub struct BasicImage<const WIDTH: usize, const HEIGHT: usize> { /* private fields */ }

This is a struct to store a raw image and can provide basic image thresholding.

Example

The following example will:

Create an image 10x5 in size.

Set pixels (0,0) = 200, (1,1) = 100, (9,4) = 10.

Create a histogram of the different pixel intensities.

Provide a threshold value which isolates the background from the foreground.

use star_tracker_lib::image_processing::{Image, BasicImage};
use star_tracker_lib::util::aliases::{UInt, Byte};
use star_tracker_lib::util::list::ArrayList;
use star_tracker_lib::util::units::Pixel;

const WIDTH : usize = 10;
const HEIGHT: usize = 5;
let mut img : BasicImage<WIDTH, HEIGHT> = BasicImage::new();     // Creates a black image.
assert_eq!(img.width(), WIDTH);     // The maximum width of the image.
assert_eq!(img.height(), HEIGHT);   // The maximum height of the image.

assert_eq!(img.get(Pixel{x: 9, y: 4}), 0);          // Every pixel defaults at 0.
img.set(Pixel{x: 9, y: 4}, 10);                     // Sets (9,4) to 10.
assert_eq!(img.get(Pixel{x: 9, y: 4}), 10);         // The pixel has been changed to 10.
img.set(Pixel{x: 1, y: 1}, 100);                    // Set pixel (x: 1, y: 1) to 100.
img.set(Pixel{x: 0, y: 0}, 200);                    // Set pixel (x: 0, y: 0) to 200.
img.set(Pixel{x: 0, y: 1}, 200);                    // Set pixel (x: 0, y: 1) to 200.

// Generate Histogram
let mut hist : [UInt; Byte::max_value() as usize + 1] = [0; Byte::MAX as usize + 1];   // Can be of any size from 0 to the max intensity.
img.histogram(&mut hist);
assert_eq!(hist[0], (HEIGHT * WIDTH) as UInt - 4);   // Everything defaults to 0.
assert_eq!(hist[10], 1);                             // intensity 10 has 1 pixel.
assert_eq!(hist[100], 1);                            // 1 pixel is at intensity of 100.
assert_eq!(hist[200], 2);                            // 2 pixels are at the intensity of 200.

Implementations

impl<const WIDTH: usize, const HEIGHT: usize> BasicImage<WIDTH, HEIGHT>

pub fn new() -> BasicImage<WIDTH, HEIGHT>

Creates an image with the default value of 0.

Example

use star_tracker_lib::image_processing::{BasicImage, Image};
use star_tracker_lib::util::units::Pixel;
let img : BasicImage<1, 1> = BasicImage::new();
assert_eq!(img.get(Pixel{x: 0, y: 0}), 0);