Struct star_tracker_lib::image_processing::Blob

pub struct Blob {
    pub size: usize,
    pub intensity: UInt,
    pub centroid: Vector2,
}

A set of connected pixels in the foreground representing a star.

This uses the grass fire method (iteratively find new pixels and consume them).

Example

use star_tracker_lib::image_processing::Blob;
use star_tracker_lib::image_processing::BasicImage;
use star_tracker_lib::image_processing::Image;
use star_tracker_lib::image_processing::ThresholdPercent;			
use star_tracker_lib::util::units::Pixel;
use star_tracker_lib::util::list::{List, ArrayList};
let mut img : BasicImage<3, 3> = BasicImage::new();
img.set(Pixel{x: 0, y: 0}, 1); // 1 1 0
img.set(Pixel{x: 1, y: 0}, 1); // 1 0 0
img.set(Pixel{x: 0, y: 1}, 1); // 0 0 1
img.set(Pixel{x: 2, y: 2}, 1);

// lst is all the blobs in the image.
let mut lst : ArrayList<Blob, 9> = ArrayList::new();

// stack is the memory allocation allowed for when finding the blob.
// If you have limited space, use an array list with the size equal to the max number of pixels in a blob.
// If you have unlimited space or a heap, use Vec.
const MAX_BLOB_SIZE : usize = 3;
let mut stack_on_stack: ArrayList<Pixel, MAX_BLOB_SIZE> = ArrayList::new();
// let mut stack_on_heap : Vec<Pixel> = Vec::new();

// You must have a threshold for the image to know what is foreground.
let thresh = ThresholdPercent::new(&img, 0.5); // Look at ThresholdGrid for a better threshold.

// The smallest size a blob is allowed.
let min_size = 2; 

Blob::find_blobs(min_size, &thresh, &mut img, &mut stack_on_stack, &mut lst);
assert_eq!(lst.get(0).intensity, 3);
assert_eq!(lst.get(0).centroid.x, 0.333333333333333333333);
assert_eq!(lst.get(0).centroid.y, 0.333333333333333333333);

assert_eq!(lst.size(), 1);

// Image has been consumed.
assert_eq!(img.get(Pixel{x: 0, y: 0}), 0);
assert_eq!(img.get(Pixel{x: 1, y: 0}), 0);
assert_eq!(img.get(Pixel{x: 0, y: 1}), 0);
assert_eq!(img.get(Pixel{x: 2, y: 2}), 0);

Fields
size: usize

The number of pixels used.

intensity: UInt

The combined intensity of all the pixels.

centroid: Vector2

The center weighted point.

Implementations

impl Blob

pub fn new() -> Blob

pub fn find_blobs( min_size: usize, threshold: &dyn Threshold, img: &mut dyn Image, stack: &mut dyn List<Pixel>, lst: &mut dyn List<Blob> )

Finds all blobs in an image.

Arguments

min_size - The least pixels to make a blob.

threshold - The threshold of the image.

img - The image to read and set to black.

stack - Used to store pixels of the blob, either use a Vec if you dont care about the size of an array list if the blob cannot exceed a certain size.

lst - The list to append the blobs to.

Example

use star_tracker_lib::image_processing::Blob;
use star_tracker_lib::image_processing::BasicImage;
use star_tracker_lib::image_processing::Image;
use star_tracker_lib::image_processing::ThresholdPercent;			
use star_tracker_lib::util::units::Pixel;
use star_tracker_lib::util::list::{List, ArrayList};
let mut img : BasicImage<3, 3> = BasicImage::new();
img.set(Pixel{x: 0, y: 0}, 1); // 1 1 0
img.set(Pixel{x: 1, y: 0}, 1); // 1 0 0
img.set(Pixel{x: 0, y: 1}, 1); // 0 0 1
img.set(Pixel{x: 2, y: 2}, 1);

// lst is all the blobs in the image.
let mut lst : ArrayList<Blob, 9> = ArrayList::new();

// stack is the memory allocation allowed for when finding the blob.
// If you have limited space, use an array list with the size equal to the max number of pixels in a blob.
// If you have unlimited space or a heap, use Vec.
const MAX_BLOB_SIZE : usize = 3;
let mut stack_on_stack: ArrayList<Pixel, MAX_BLOB_SIZE> = ArrayList::new();
// let mut stack_on_heap : Vec<Pixel> = Vec::new();

// You must have a threshold for the image to know what is foreground.
let thresh = ThresholdPercent::new(&img, 0.5);

// Blobs must be equal to or above this size to added.
let min_size = 2;

Blob::find_blobs(min_size, &thresh, &mut img, &mut stack_on_stack, &mut lst);
assert_eq!(lst.get(0).intensity, 3);
assert_eq!(lst.get(0).centroid.x, 0.333333333333333333333);
assert_eq!(lst.get(0).centroid.y, 0.333333333333333333333);

assert_eq!(lst.size(), 1);

// Image has been consumed.
assert_eq!(img.get(Pixel{x: 0, y: 0}), 0);
assert_eq!(img.get(Pixel{x: 1, y: 0}), 0);
assert_eq!(img.get(Pixel{x: 0, y: 1}), 0);
assert_eq!(img.get(Pixel{x: 2, y: 2}), 0);

pub fn spread_grass_fire( threshold: &dyn Threshold, start: Pixel, img: &mut dyn Image, stack: &mut dyn List<Pixel> ) -> Blob

Creates and expands the blob to fit any pixels in the foreground.

Arguments

threshold - The value which is the minimum intensity to satisfy being in the foreground.

start - The start point where the blob should begin.

img - The image to read and consume (set pixels to 0).

stack - Used to store pixels of the blob, either use a Vec if you dont care about the size of an array list if the blob cannot exceed a certain size.

Returns

The blob.

Example

use star_tracker_lib::image_processing::{Blob, BasicImage, Image, ThresholdPercent};
use star_tracker_lib::util::{list::{List, ArrayList}, units::Pixel};
let mut img : BasicImage<2, 2> = BasicImage::new();
img.set(Pixel{x: 0, y: 0}, 1); // 1 1
img.set(Pixel{x: 1, y: 0}, 1); // 1 0
img.set(Pixel{x: 0, y: 1}, 1);

// stack is the memory allocation allowed for when finding the blob.
// If you have limited space, use an array list with the size equal to the max number of pixels in a blob.
// If you have unlimited space or a heap, use Vec.
const MAX_BLOB_SIZE : usize = 3;
let mut stack_on_stack: ArrayList<Pixel, MAX_BLOB_SIZE> = ArrayList::new();
// let mut stack_on_heap : Vec<Pixel> = Vec::new();

// You must have a threshold for the image to know what is foreground.
let thresh = ThresholdPercent::new(&img, 0.5);

let start : Pixel = Pixel{x: 0, y: 0};
let blob = Blob::spread_grass_fire(&thresh, start, &mut img, &mut stack_on_stack);
assert_eq!(blob.intensity, 3);

assert!((blob.centroid.x - 0.333).abs() < 0.01);
assert!((blob.centroid.y - 0.333).abs() < 0.01);

assert_eq!(img.get(Pixel{x: 0, y: 0}), 0);
assert_eq!(img.get(Pixel{x: 1, y: 0}), 0);
assert_eq!(img.get(Pixel{x: 0, y: 1}), 0);

pub fn find_neighbours( threshold: &dyn Threshold, pt: &Pixel, img: &dyn Image, stack: &mut dyn List<Pixel> )

Finds all neighbouring pixels which are valid and within the theshold.

Arguments

threshold - The value which is the minimum intensity to satisfy being in the foreground.

pt - The point to look around.

img - The image to examine.

stack - The stack to append to.

Example

use star_tracker_lib::image_processing::{Blob, BasicImage, Image, ThresholdPercent};
use star_tracker_lib::util::units::Pixel;
use star_tracker_lib::util::list::{ArrayList, List};
let threshold = 1;
let pt : Pixel = Pixel{ x: 1, y: 1 };
let mut img : BasicImage<3, 3> = BasicImage::new();
let mut lst : ArrayList<Pixel, 4> = ArrayList::new();

img.set(Pixel{x: 1, y: 0}, 1); // 0, 1, 0
img.set(Pixel{x: 0, y: 1}, 1); // 1, 0, 1
img.set(Pixel{x: 1, y: 2}, 1); // 0, 1, 0
img.set(Pixel{x: 2, y: 1}, 1);

	
// You must have a threshold for the image to know what is foreground.
let thresh = ThresholdPercent::new(&img, 0.5);

Blob::find_neighbours(&thresh, &pt, &img, &mut lst); // Right, Left, Up, Down
assert_eq!(lst.size(), 4);

assert_eq!(lst.get(0).x, 2); // Right
assert_eq!(lst.get(0).y, 1);

assert_eq!(lst.get(1).x, 0); // Left
assert_eq!(lst.get(1).y, 1);

assert_eq!(lst.get(2).x, 1); // Up
assert_eq!(lst.get(2).y, 0);

assert_eq!(lst.get(3).x, 1); // Down
assert_eq!(lst.get(3).y, 2);

pub fn find_centroid( b_pos: Decimal, b_intensity: UInt, p_pos: UInt, p_intensity: UInt ) -> Decimal

Finds the new centroid after the pixel is added to the blob on a specific axis.

Arguments

b_pos - The current blob centroid.

b_intensity - The magnitude of the pixels in the blob.

p_pos - The pixel position.

p_intensity - The magnitude of the pixel.

Returns

The new centroid on the given axis.

Example

use star_tracker_lib::image_processing::Blob;
assert_eq!(Blob::find_centroid(0.0, 3, 1, 1), 0.25); // 2, 1
assert_eq!(Blob::find_centroid(0.0, 9, 2, 1), 0.2); // 9, 0, 1

pub fn to_vector2(blobs: &dyn List<Blob>, points: &mut dyn List<Vector2>)

Converts the list of blobs to a list of points.

Arguments

blobs - The blobs to convert to points.

points - The centroid of the blobs.

Example

use star_tracker_lib::util::{aliases::Decimal, units::Vector2, list::{List, ArrayList}};
use star_tracker_lib::image_processing::Blob;
let mut blobs : ArrayList<Blob, 3> = ArrayList::new();
blobs.push_back(Blob{size: 0, intensity: 10, centroid: Vector2{x: 10.0, y: 10.0}});
blobs.push_back(Blob{size: 0, intensity: 5, centroid: Vector2{x: 5.0, y: 5.0}});
blobs.push_back(Blob{size: 0, intensity: 0, centroid: Vector2{x: 0.0, y: 0.0}});

let mut points : ArrayList<Vector2, 2> = ArrayList::new();
Blob::to_vector2(&blobs, &mut points);
assert_eq!(blobs.get(0).centroid, points.get(0));
assert_eq!(blobs.get(1).centroid, points.get(1));

pub fn sort_descending_intensity(brightest: &Blob, dullest: &Blob) -> bool

Sorts in order of intensity brightest first.

Arguments

brightest - The brightest value.

dullest - The least intense blob.

Return

True if in order.

Example

use star_tracker_lib::image_processing::Blob;
use star_tracker_lib::util::units::Vector2;
let brightest = Blob{size: 0, intensity: 1, centroid: Vector2{x: 0.0, y: 0.0}};
let dullest = Blob{size: 0, intensity: 0, centroid: Vector2{x: 0.0, y: 0.0}};
assert!(Blob::sort_descending_intensity(&brightest, &dullest));
assert!(!Blob::sort_descending_intensity(&dullest, &brightest));

pub fn sort_descending_size_intensity(largest: &Blob, dullest: &Blob) -> bool

Sorts considering the brightness and the size of the blob. This helps nullify dead pixels or inaccurate anomalies.