Struct star_tracker_lib::image_processing::ThresholdPercent

pub struct ThresholdPercent { /* private fields */ }

A basic percent threshold (Worse Than ThresholdGrid).

This will generate a number based on a percentage brightness of the image.

Implementations

impl ThresholdPercent

pub fn new(img: &dyn Image, percent: Decimal) -> Self

Generates a threshold for the given image.

Arguments

img - The image

percent- What percentage of the darkest pixels should be ignored? (0.9999).

Returns

Returns threshold with the given percent.

Example

use star_tracker_lib::image_processing::ThresholdPercent;
use star_tracker_lib::image_processing::Threshold;
use star_tracker_lib::image_processing::BasicImage;
use star_tracker_lib::image_processing::Image;
use star_tracker_lib::util::units::Pixel;
let mut img : BasicImage<3, 3> = BasicImage::new();
img.set(Pixel{x: 0, y: 0}, 1); img.set(Pixel{x: 1, y: 0}, 4); img.set(Pixel{x: 2, y: 0}, 5);
img.set(Pixel{x: 0, y: 1}, 2); img.set(Pixel{x: 1, y: 1}, 4); img.set(Pixel{x: 2, y: 1}, 6);
img.set(Pixel{x: 0, y: 2}, 3); img.set(Pixel{x: 1, y: 2}, 4); img.set(Pixel{x: 2, y: 2}, 7);
// Each pixel is 0.12%

let thresh = ThresholdPercent::new(&img, 0.11);
assert_eq!(thresh.foreground(Pixel{x: 0, y: 0}), 1);
let thresh = ThresholdPercent::new(&img, 0.12);
assert_eq!(thresh.foreground(Pixel{x: 0, y: 0}), 2);
let thresh = ThresholdPercent::new(&img, 0.24);
assert_eq!(thresh.foreground(Pixel{x: 0, y: 0}), 3);
 

pub fn threshold( img: &dyn Image, percentage: Decimal, histogram: &[UInt] ) -> Byte

Finds the minimum bar of the histogram which satisfies the percentage of pixels.

Arguments

percentage - The minimum % of pixels to be foreground.

histogram - The histogram to examine.

Returns

The minimum brightness value to be foreground.

Example

use star_tracker_lib::image_processing::BasicImage;
use star_tracker_lib::image_processing::Image;
use star_tracker_lib::image_processing::ThresholdPercent;
use star_tracker_lib::util::list::ArrayList;
use star_tracker_lib::util::aliases::UInt;
use star_tracker_lib::util::aliases::Byte;
use star_tracker_lib::util::units::Pixel;

let img : BasicImage<16, 16> = BasicImage::new();
let hist :  [UInt; Byte::max_value() as usize + 1] = [1; Byte::MAX as usize + 1]; // [1, 1, ...]
assert_eq!(ThresholdPercent::threshold(&img, 0.5, &hist), Byte::MAX / 2 + 1);
assert_eq!(ThresholdPercent::threshold(&img, 0.0, &hist), 0);
assert_eq!(ThresholdPercent::threshold(&img, 1.0, &hist), Byte::MAX);