Module star_tracker_lib::tracking_mode

tracking_mode is the method of matching the stars found in the image to a database.

This program uses the pyramid method of star tracking.

Pyramid Tracking Mode Algorithm

The pyramid method is one of the most reliable methods. It was developed by Mortari in 1997.
The method was modified a few times.

The algorithm creates a kernel (a set of 3 stars) with a method to reduce the number of times a false positive star is used.
It then will compare the constellation with a k-vector table database.
If it succeeds, it will then find another star and verify it is also consistent.

Creating the kernel

To select the stars, a looping algorithm must be chosen the proposed method is implemented in this algorithm.
This will access the index of 0-1-2, 2-3-4, 3-4-5, 1-2-4, 2-3-5, 1-2-5, 1-3-4, 2-4-5, 1-3-5, 1-4-5.
This method is done to reduce the times a specific star is used as it may be a false star.

Using the kernel

By working with the database, a set of star ‘pairs’ are found.
These each pair contains the id/location of each star.
By finding the common star from 2 pairs, the star can be identified in the database and the sample.
The matched triangle will then have its specularity checked (check if it is flipped).

Pyramid

If a correct kernel is found, an extra test is performed where another star is picked, finding the distance from this star to the other stars, this is then compared with the database.

Database

Refer to this file.

Features

This method features:

An algorithm which can fit on a microcontroller.

Contains a method to reduce the repetition of using a false star.

Fast as uses K-Table.

Checks if the triangle is flipped (false positive).

Adds an extra star to verify correct identification opposing other triangle constellation based methods.

Useful links:

Original Paper

Pseudo Interpretation

Good Explanation

How to Use

use star_tracker_lib::tracking_mode::database::MockDatabase;
use star_tracker_lib::tracking_mode::database::Database;
use star_tracker_lib::tracking_mode::AbandonSearchFailures;
use star_tracker_lib::tracking_mode::StarTriangleIterator;
use star_tracker_lib::tracking_mode::ConstellationResult;
use star_tracker_lib::tracking_mode::AbandonSearch;
use star_tracker_lib::tracking_mode::Constellation;
use star_tracker_lib::tracking_mode::Specularity;

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Degrees;
use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Match;
use star_tracker_lib::util::list::List;

// To construct the database, use the tools in *star_tracker_nix*, look at demo.
let database_iterator = // database goes here.

// Pretend these are image values which have been projected onto a unit sphere.
let stars_3d: Vec<Equatorial> = vec![];

// This is a way for increasing the speed of the algorithm.
// Just use an array size which is large and you can fit...
const SOME_LARGE_ARRAY_SIZE: usize = 1000;
let iterator = StarTriangleIterator::<SOME_LARGE_ARRAY_SIZE>::new();

// The algorithm tries to create a triangle with a set of stars. 
// If the triangle is flipped, it is not a proper match.
// Specularity is a way of doing this.
// There is a min specularity where it is effectively a strait line.
// In this case, if the specularity is below this number, specularity is ignored.
// Specularity::default() has a good value, however if this isn't working for you, have a play around.
// Look in *Specularity* for more details.
let specularity = Specularity::default();

// This will abort the program if it is taking too long.
// This abort uses the number of fails.
// The algorithm will try and construct a triangle with 3 stars, if it cannot, a fail occurs.
// This is set so if it cannot make a triangle 10 times in a row, it will abort.
// It is important to abort early as a success usually takes less than 5 times.
// The more fails, the more likely it will be incorrect and unreliable.
// You should also team this up with a timer, however this is not platform specific so make your own.
// There is a timer and fails counter abort in *star_tracker_nix* used by *demo*. 
let fails_good = 10;   // Unsuccessful matches.
let mut abort = AbandonSearchFailures::new(fails_good);

// How inaccurate is the lens?
// You will need to test the lens.
// Use nova.astrometry.net and get the corr file, put it through corr_analyzer in star_tracker_nix.
// Don't use an error above 0.1 degrees as it will likely fail.
let angle_tolerance = Degrees(0.08).to_radians();

// The min and max (inclusive) required stars.
// If not enough stars can be formed into a constellation, the search will continue and will eventually fail.
// If there is too many stars in the constellation, the algorithm will end early on the upper bounds to save time.
// The pyramid method should use 4 stars, so it is set to 4.
let required_stars = 4..=4;

let mut output: Vec<Match<usize>> = Vec::new();

	let success = Constellation::find_all (
	&stars_3d, &mut database_iterator, &mut iterator,
	&mut specularity, &mut abort, angle_tolerance, required_stars, &mut output 
);

match success
{
	ConstellationResult::ErrorNoTriangleMatch      { fails } => 
		println!("FAILED: Could not match any stars; {} failures.", fails),

	ConstellationResult::ErrorAborted              { fails } =>
		println!("FAILED: Aborted due to AbandonSearch parameter; {} failures.", fails),

	ConstellationResult::ErrorInsufficientPyramids { fails } =>
		println!("FAILED: Not enough matched stars; {} failures.", fails),

	ConstellationResult::Success                   { fails } =>
		println!("SUCCESS; with {} fails.", fails),
}

Modules

constellation

Implementation of Constellation.

database

The database for the pyramid method is a set of 2 lookup tables (star pairs, star catalogue) and an equation (k-vector)

kernel_iterator

Implementation for KernelIterator

pilot_finder

Implementation of StarPyramid.

specularity

Implementation for Specularity

star_pair

Implementation for KernelIterator

star_triangle

Implementation of StarTriangle

star_triangle_iterator

Structs

AbandonSearchFailures

A type of AbandonSearch

KernelIterator

An iterator which reduces the chances of getting the same false star twice.

PilotFinder

Use to search for a pilot star to match a star triangle.

Specularity

Ensures the constellation is the correct orientation.

StarPair

A set of 2 stars in equatorial space, this represents a line / angle.

StarPyramid

A set of 4 stars in 3D space, this represents a pyramid.

StarTriangle

A set of 3 stars in T space, this represents a triangle.

StarTriangleIterator

A way of speeding up the database search algorithm.

Enums

Constellation

The return type for the star pyramid.

ConstellationResult

The output from find_all.

SpecularityResult

A specularity test.

Traits

AbandonSearch

Used to tell the search that it should give up.

PyramidConstruct

Used to find the pilot star for the found star triangle.

RegionCompare

Use to compare regions.

SpecularityConstruct

Used to check the specularity of a star triangle.

TriangleConstruct

Used to iterate over star triangles until a match is found.