Module star_tracker_lib::tracking_mode::database

The database for the pyramid method is a set of 2 lookup tables (star pairs, star catalogue) and an equation (k-vector)

K-Vector

The k-vector is a fast lookup table which has a time complexity independent of the size of the database. It uses an equation to locate specific bins in the database where the element can be found.

Equations

Variables

m - Gradient

q - Intercept

y - The input variable (angular distance).

Prepossessing

This is the process of creating the k-vector, this is only required once. Symbols

D_max - The maximum value of an element in the database.

D_min - The minimum value of an element in the database.

n - The number of elements in the database.

N - The number of vector bins, a larger value will be faster, however it will cost more memory, ideal: N = n.

d - (n - 1)

e - Relative machine precision rust uses binary32/64.

f32 = 1.19e-07

f64 = 2.22e-16

Finding Gradient m = ( D_max - D_min + 2 d e ) / ( N - 1 )

Finding Intercept q = D_min - d e

Searching

By inputting 2 equations, the bounds of the element can be found. k_a = floor(y - q) / m + 1 k_b = ceil(y - q) / m i is within k_a and k_b.

List of Pairs

The list of star pairs is a lookup table containing 2 references to stars.

The stars will have their angular distances calculated, this will be the order they are sorted in.

The stars should only form a pair if they are within the same field of view.

The index and position are not required to store. e.g.

index	a	b

0	*star a	*star b

1	*star c	*star b

2	*star c	*star a

…	…	…

Star Catalogue

The star catalogue is just the positions of the stars. This can be done using equatorial (less memory) or Cartesian coordinates (less time). The equatorial is more beneficial for this project. e.g.

index (*star)	ra	dec

0 (*star a)	1	-1

1 (*star b)	2	1

2 (*star c)	2	-1

3 (*star d)	3	0

…	…	…

Re-exports

pub use crate::tracking_mode::database::database::Database;

pub use crate::tracking_mode::database::chunk_iterator::ChunkIterator;

Modules

chunk_iterator

Implementation of ChunkIterator

database

Used for any database. A set of functions may not be relevant to the required database, however, it is made universal.

pyramid_database

regional_database

search_result

Implementation of SearchResult.

Structs

ChunkAreaSearch

Like ChunkIteratorEquatorial except it does not iterate.
This is used so that you can search a small reagion of the sky.
The usecase is if you have a general idea of the location of the attitude but not good enough. For areas that go past 360 degrees, either use a negative min bounds or a value over 360 for max.

ChunkIteratorDeclination

A chunk iterator is a way of optimizing the database search process.
When searching the database for star matches, the database will return matches from around the entire celestial sphere.
To ensure that all the pairs are within the same camera frame, a chunk iterator is used to move between each chunk/region.
This ensures that only stars within the field of view will be used.

ChunkIteratorEquatorial

A chunk iterator is a way of optimizing the database search process.
When searching the database for star matches, the database will return matches from around the entire celestial sphere.
To ensure that all the pairs are within the same camera frame, a chunk iterator is used to move between each chunk/region.
This ensures that only stars within the field of view will be used.

ChunkIteratorNone

A chunk iterator is a way of optimizing the database search process.
When searching the database for star matches, the database will return matches from around the entire celestial sphere.
To ensure that all the pairs are within the same camera frame, a chunk iterator is used to move between each chunk/region.
This ensures that only stars within the field of view will be used.

ChunkIteratorRegional

A chunk iterator is a way of optimizing the database search process.
When searching the database for star matches, the database will return matches from around the entire celestial sphere.
To ensure that all the pairs are within the same camera frame, a chunk iterator is used to move between each chunk/region.
This ensures that only stars within the field of view will be used.

KVector

The database equation which points to the star pair database.

PyramidDatabase

The compiled k-vector database, this will allow the construction of the database and the lookup of elements in the database.

RegionalDatabase

This database is required for the ChunkIteratorRegional.
A list specifying the chunk the star is located in is required as the iterator uses a bitfield.

SearchResult

The result for querying for a close match star pair. Used in Database::find_close_ref.

Traits

KVectorSearch