Trait star_tracker_lib::tracking_mode::database::chunk_iterator::ChunkIterator

pub trait ChunkIterator {
    // Required methods
    fn begin(&mut self);
    fn next(&mut self) -> bool;
    fn get_database(&self) -> &dyn Database;
    fn same_region(&self, pair_index: usize) -> bool;

    // Provided method
    fn find_close_ref_region(
        &self,
        find: Radians,
        tolerance: Radians,
        found: &mut dyn List<SearchResult>
    ) { ... }
}

Required Methods

fn begin(&mut self)

Resets the iterator.
Call this to return to the start, then call next.

fn next(&mut self) -> bool

Steps to the next region. This must be called to begin. If there is no more regions, next will return false.

fn get_database(&self) -> &dyn Database

Returns the database this iterator holds.

fn same_region(&self, pair_index: usize) -> bool

Returns true if the pair at the provided index fits the current region.

Provided Methods

fn find_close_ref_region( &self, find: Radians, tolerance: Radians, found: &mut dyn List<SearchResult> )

Finds close matches to the provided angular separation and returns the star pair reference. The an element of the star pair reference can be inserted into find_star to get the actual location.

Arguments

find - The angular separation between the found stars to find in the database.

found - The closest matches to the provided find.

Returns

The fields at the end of each pair. Use this information so that when finding multiple pairs connected, they can span multiple regions.