Trait star_tracker_lib::tracking_mode::database::database::Database

pub trait Database {

    // Required methods
    fn get_pairs(&self, index: usize) -> StarPair<usize>;
    fn get_pairs_size(&self) -> usize;
    fn get_catalogue(&self, index: usize) -> Equatorial;
    fn get_catalogue_size(&self) -> usize;
    fn get_k_vector(&self, index: usize) -> usize;
    fn get_k_vector_size(&self) -> usize;
    fn get_k_lookup(&self) -> KVector;
    fn get_fov(&self) -> Radians;

    // Provided methods
    fn find_star(&self, index: usize) -> Error<Equatorial> { ... }
    fn find_close_ref(
        &self,
        find: Radians,
        tolerance: Radians,
        found: &mut dyn List<SearchResult>
    ) { ... }
    fn find_close_ref_range(
        &self,
        find: Radians,
        tolerance: Radians
    ) -> Range<usize>  { ... }
    fn trim_range(
        &self,
        find: Radians,
        tolerance: Radians,
        range: Range<usize>
    ) -> Range<usize>  { ... }
    fn angle_distance(&self, pair: StarPair<usize>) -> Error<Radians> { ... }

}

Required Methods

fn get_pairs(&self, index: usize) -> StarPair<usize>

Gets the star pair at the index in the array. Used for any trait implementations bellow.

fn get_pairs_size(&self) -> usize

Gets the size of the star pairs array. Used for any trait implementations bellow.

fn get_catalogue(&self, index: usize) -> Equatorial

Gets the catalogue array. Used for any trait implementations below.

fn get_catalogue_size(&self) -> usize

Gets the catalogue array. Used for any trait implementations bellow.

fn get_k_vector(&self, index: usize) -> usize

Gets the k_vector array. Used for any trait implementations bellow.

fn get_k_vector_size(&self) -> usize

Gets the k_vector array. Used for any trait implementations bellow.

fn get_k_lookup(&self) -> KVector

Gets the k_vector lookup equation. Used for any trait implementations below.

fn get_fov(&self) -> Radians

Gets the field of view the database was created for.

Provided Methods

fn find_star(&self, index: usize) -> Error<Equatorial>

Finds the star with the provided index from the star pair database.

Arguments

index - The index of the star in the catalogue database.

Returns

The actual position (usually J2000).

fn find_close_ref( &self, find: Radians, tolerance: Radians, found: &mut dyn List<SearchResult> )

Finds close matches to the provided angular separation and returns the star pair reference. The an element of the star pair reference can be inserted into find_star to get the actual location.

Arguments

find - The angular separation between the found stars to find in the database.

found - The closest matches to the provided find.

fn find_close_ref_range( &self, find: Radians, tolerance: Radians ) -> Range<usize>

Finds the range in the pair list where pairs of the given distance will be.
This is different to find_close_ref as it allows you to manually add pairs to the list.
This helps if you need a range.

find - The angular separation between the found stars to find in the database.

found - The closest matches to the provided find.

fn trim_range( &self, find: Radians, tolerance: Radians, range: Range<usize> ) -> Range<usize>

Trims the range provided by the k-vector so that every value is within the tolerance.

Arguments

find - The center point of the tolerance.

tolerance - The distance allowed from the center.

range - The range to be trimmed.

Returns

The a trimmed version of range.

fn angle_distance(&self, pair: StarPair<usize>) -> Error<Radians>

Finds the angular distance between a star pair referencing the catalogue.

Arguments

pair - The pair to find the distance from.

Returns

The angular distance between the pair.