Struct star_tracker_lib::tracking_mode::database::KVector

pub struct KVector {
    pub gradient: Decimal,
    pub intercept: Decimal,
    pub min_value: Radians,
    pub max_value: Radians,
    pub num_bins: usize,
}

The database equation which points to the star pair database.

Fields
gradient: Decimal

The gradient for the k-vector bin equation. This also represents ~ 1/2 the tolerance or the distance between bins.

intercept: Decimal

The y intercept for the k-vector bin equation.

min_value: Radians

The smallest value in the database.

max_value: Radians

The largest value in the database.

num_bins: usize

The number of kvector elements.

Implementations

impl KVector

pub fn new(num_bins: usize, min_value: Decimal, max_value: Decimal) -> KVector

For Preprocessing.
Finds the values required to use a K-Vector with the given database.

Arguments

num_bins - The number of bins.

sorted_database - A database in increasing sorted order by the angular distance between the stars.

Returns

A calibrated KVector to use for the current configuration.

Example

use star_tracker_lib::tracking_mode::database::KVector;
use star_tracker_lib::util::test::DECIMAL_PRECISION_TEST;
use star_tracker_lib::util::test::TestEqual;
use star_tracker_lib::tracking_mode::StarPair;
use star_tracker_lib::util::aliases::Decimal;
use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;

let element_min = 1.23;
let element_max = 10.0;
let num_bins = 2;
let k_vector = KVector::new(num_bins, element_min, element_max);
 
let machine_epsilon = DECIMAL_PRECISION_TEST;
let gradient = (element_max - element_min + machine_epsilon * 2.0) / (num_bins as Decimal);
let intercept = element_min - machine_epsilon;
// assert!(k_vector.gradient.test_equal(&(gradient as Decimal)));
// assert!(k_vector.intercept.test_equal(&(intercept as Decimal)));
// assert_eq!(k_vector.min_value, Radians(element_min as Decimal));
// assert_eq!(k_vector.max_value, Radians(element_max as Decimal));
 
let mut y = gradient * 0.0 + intercept;			// The lower bounds.
 
 
assert!(y < element_min);						// Must be smaller than min value.
assert!(element_min - y < 0.05);				// Must be close to the min value.
 
y = gradient * 1.0 + intercept;					// This is the middle bounds.
assert!(element_min < y && y < element_max);	// Value must be greater than the smallest element.

// Must be in the center between the bounds.
assert!((y - (element_max + element_min) / 2.0).abs() < 0.0001);
 
y = gradient * 2.0 + intercept;					// This is the upper bounds.
assert!(element_max <= y);						// Must include all values.
assert!(y - element_max < 0.05);				// Must be close to the max value.