Struct star_tracker_lib::tracking_mode::database::RegionalDatabase

pub struct RegionalDatabase<'a> {
    pub fov: Radians,
    pub num_fields: UInt,
    pub k_lookup: KVector,
    pub k_vector: &'a dyn LinearLookup<usize>,
    pub pairs: &'a dyn LinearLookup<StarPair<usize>>,
    pub catalogue: &'a dyn LinearLookup<Equatorial>,
    pub catalogue_field: &'a dyn LinearLookup<BitField>,
}

This database is required for the ChunkIteratorRegional.
A list specifying the chunk the star is located in is required as the iterator uses a bitfield.

Fields
fov: Radians

The field of view used when the database was constructed.

num_fields: UInt

The number of bits in the catalogue_field bit field.

k_lookup: KVector

The equation of the k_vector which points to the k_vector table k_vector.

k_vector: &'a dyn LinearLookup<usize>

The k_vector table pointed to from k_lookup.

pairs: &'a dyn LinearLookup<StarPair<usize>>

Pairs of stars in order of separation.
The separation is not provided, use the k_lookup, k_vector to find stars with a specific separation.
This points to the cataloge, where each element in the star pair is an index of the catalogue.

catalogue: &'a dyn LinearLookup<Equatorial>

The stars location in the sky. Use the k_lookup, k_vector, pairs to find the pairs of stars.

catalogue_field: &'a dyn LinearLookup<BitField>

Reflects the index of the catalogue.
This represents what region/chunk of the sky the field is located in. Use this to identify if a star is relevant.