Enum star_tracker_lib::tracking_mode::Constellation

pub enum Constellation {
    Pyramid(Match<StarPyramid<Equatorial>>),
    Triangle(Match<StarTriangle<Equatorial>>),
    None,
}

The return type for the star pyramid.

Either there is no match or less than 3 stars (None)
There is a match but only 3 supplied stars (Triangle)
There is a match and more than 4 supplied stars (Pyramid)

Variants

Pyramid(Match<StarPyramid<Equatorial>>)

Triangle(Match<StarTriangle<Equatorial>>)

None

Implementations

impl Constellation

pub fn find( stars: &dyn List<Equatorial>, database: &mut dyn ChunkIterator, gen_tri: &mut dyn TriangleConstruct, gen_pyr: &mut dyn PyramidConstruct, gen_spec: &mut dyn SpecularityConstruct, angle_tolerance: Radians, abort: &mut dyn AbandonSearch ) -> Constellation

Finds a StarPyramid or StarTriangle if possible.

Arguments

stars - The observed (image) stars.

database - The database storing reference

gen_tri - An object for generating a StarTriangle.

gen_pyr - An object for generating a Pyramid.

gen_spec - An object for checking Specularity.

angle_tolerance - How much error a star pair can have until it is not considered the same. Used for searching the database.

abort - A way of abandoning a search if it takes too long.

Returns

If a star pyramid can be formed: Match{input: Observed Stars, output: Corresponding Database}. None if a star pyramid and star triangle cannot be formed.

NOT IMPLEMENTED (in the case of a star triangle, it is not as accurate, left out) If a star triangle can be formed: Match{input: Observed Stars, output: Corresponding Database}.

pub fn find_all( stars: &dyn List<Equatorial>, database: &mut dyn ChunkIterator, gen_tri: &mut dyn TriangleConstruct, gen_spec: &mut dyn SpecularityConstruct, abort: &mut dyn AbandonSearch, angle_tolerance: Radians, num_stars: RangeInclusive<usize>, matches: &mut dyn List<Match<usize>> ) -> ConstellationResult

Finds as many star matches as possible until required is met. Similar to find except it uses indices and allows you to increase the accuracy past 4 stars. If in the current iteration, more than required is found, all of the found stars will be used.

Arguments

stars - The observed (image) stars.

database - The database storing reference

gen_tri - An object for generating a StarTriangle.

gen_pyr - An object for generating a Pyramid.

gen_spec - An object for checking Specularity.

angle_tolerance - How much error a star pair can have until it is not considered the same. Used for searching the database.

num_stars - The how many stars are required.
If min is reached, the current constellation will be returned.
If between min and max stars, more stars will be searched for, on max, the constellation is immediately returned and the search ends.
If min is reached, true is returned, otherwise it will continue searching.
It is recommended to use 4..=4 for 4 stars to be found.

abort - A way of abandoning a search if it takes too long. Ensure abort has been reset and started before calling this method.

Returns

True if the minimum number of stars could be found. False if it failed to find the minimum stars. matches will have a value if 3 stars are found, even if min stars were not found.