Struct star_tracker_lib::tracking_mode::Specularity

pub struct Specularity { /* private fields */ }

Ensures the constellation is the correct orientation.

When comparing constellation, triangles are used.
Specularity is the test to see if the triangle is flipped.
If the triangle is flipped, it is invalid.

Implementations

impl Specularity

pub fn default() -> Self

Sets up the specularity test with an ideal value.
The value will ensure all triangles which are not close to a strait line are valid.
The ideal value used is 0.001.

pub fn new(specularity_min: Decimal) -> Self

You should use Specularity::default() if you don’t know what your doing.
Specularity Min is the value until specularity should be ignored as the star triangle is a straight line.
The default value is 0.001 which has been hand picked.
Specularity ranges from ~0.0001 to ~0.01, if you use 1.0, the specularity step will be ignored.

pub fn test(&self, triangle: &StarTriangle<Vector3>) -> SpecularityResult

Finds the specularity of the triangle.

Arguments

triangle - The triangle.

specularity_min - The minimum specularity until the triangle is too small to measure.

TODO

A triangle of small area (points close together or in line) may be invalid. Any error may cause a change in specularity. Area should be considered in this function at some point.

Returns

true if +’ve specularity.

Example

use star_tracker_lib::tracking_mode::StarTriangle;
use star_tracker_lib::tracking_mode::Specularity;
use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::aliases::Decimal;

let spec = Specularity::new(239.0);

let pt1 = Vector3 { x: -1.0, y: 2.0, z: 3.0 };
let pt2 = Vector3 { x: 4.0, y: 5.0, z: 6.0 };
let pt3 = Vector3 { x: 7.0, y: 8.0, z: -9.0 };

// pt2 x pt3 = (-93, 78, -3) . pt1 = 240 = true
let st1 : StarTriangle<Vector3> = StarTriangle(pt1.clone(), pt2.clone(), pt3.clone() );
// pt2 x pt3 = (93, -78, 3) . pt1 = -240 = false
let st2 : StarTriangle<Vector3> = StarTriangle(pt1.clone(), pt3.clone(), pt2.clone() );
// pt1 x pt3 = (-42, 12, -22) . pt2 = -240 = false
let st3 : StarTriangle<Vector3> = StarTriangle(pt2.clone(), pt1.clone(), pt3.clone() );

spec.test(&st1); // SpecularityResult::Valid(true);
spec.test(&st2); // SpecularityResult::Valid(false);
spec.test(&st3); // SpecularityResult::Valid(false);