Trait star_tracker_lib::tracking_mode::PyramidConstruct

pub trait PyramidConstruct {
    // Required method
    fn find_pilot(
        &mut self,
        stars: &dyn List<Equatorial>,
        database: &dyn ChunkIterator,
        angle_tolerance: Radians,
        input: StarTriangle<usize>,
        output: StarTriangle<usize>
    ) -> Error<Match<usize>>;
}

Used to find the pilot star for the found star triangle.

Required Methods

fn find_pilot( &mut self, stars: &dyn List<Equatorial>, database: &dyn ChunkIterator, angle_tolerance: Radians, input: StarTriangle<usize>, output: StarTriangle<usize> ) -> Error<Match<usize>>

Finds the pilot

Arguments

stars - The stars from the image.

database - The database to lookup.

angle_tolerance - How much error a star pair can have until it is not considered the same. Used for searching the database.

input - The star triangle from the input (what stars are being used).

output - The star triangle from the output in the same order as input.

Returns

Ok(pilot) if valid.