Module star_tracker_lib::util

util is contains any tools which may be important for general use in this project.

Everything inside this module is expected to be used in multiple modules.

This includes;

Types of storage lists.

Aliases of datatypes to decide precision.

Coordinate systems and units that will be used.

Types of errors.

Methods for testing.

Re-exports

pub use crate::util::maths::Maths;

Modules

aliases

This module contains any alias variables to make more sense for the project.

distribution

This module is used to distribute points evenly over a sphere + some stats.

err

This provides specifics of errors.

linear_lookup

Designed to share similarities between ArrayList, Arrays and Vec.

list

A universal implementation of an array list, this is setup for a static ArrayList and Vec.

maths

A wrapper for libm.
This is needed as libm is not included by default into the no_std package.

test

This is a series of functions and traits to aid in testing.

units

This is a set of units and coordinates which are used throught the project.