Struct star_tracker_lib::util::distribution::Distribute

pub struct Distribute();

Distributes a set of points onto a sphere.

Implementations

impl Distribute

pub fn fibonacci_lattice(num_points: usize) -> Vec<Equatorial>

Sets the points of the input equatorial array as a set of close to evenly spaced points. This uses the Fibonacci golden ratio algorithm.

Arguments

num_points - The number of points on the sphere.

Example

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::aliases::Decimal;
use star_tracker_lib::util::aliases::M_PI;
use star_tracker_lib::util::distribution::Distribute;

let points = Distribute::angle_to_points(Radians(10.0));
let mut eq : Vec<Equatorial> = Distribute::fibonacci_lattice(points);

pub fn separation(points: &Vec<Equatorial>) -> Distribution

Finds the average/min/max/std arc angle between the provided points.
Use this to find how distribute the stars are from each other.
This does not take into consideration if all the stars are near each other and there is a hole elsewhere.

Use this to identify how clustered stars are from each other. It is recommended that you look at coverage.

pub fn coverage(points: &Vec<Equatorial>) -> Distribution

From choosing a random point on the unit sphere, this finds the average/min/max/std arc angle to a provided point.
Use this to ensure that there will be enough stars with a valid magnitude within the field of view of your lens.
This uses the Monti-Carlo simulation, running 10,000 times.

pub fn angle_to_points(angle: Radians) -> usize

Finds the number of points required to achieve the given angle With Error.
This is used for the fibonacci golden sphere.
The error is:

* 43deg < angle           : 0
* 24.6deg < angle < 43deg : -1
* angle < 24.6deg         : slowly increasing positive number.

As the angle gets smaller, the number of points error becomes higher, however it is insignificant.
When the error is negative, the angle is larger than it should be.

pub fn points_to_angle(points: usize) -> Radians

Finds the angle between each point given the the number of points to generate.
This uses the fibonacci golden sphere.
The error is:

* points < 12: +-2%
* 12 < points: +-2%