Trait star_tracker_lib::util::linear_lookup::LinearLookup

pub trait LinearLookup<T> {
    // Required methods
    fn size(&self) -> usize;
    fn get(&self, index: usize) -> T;
    fn set(&mut self, index: usize, val: T) -> Error<()>;
}

Designed to share similarities between ArrayList, Vec’s and Arrays.
(also works with Vec if you compile with nix feature).
When constructing the database, it is ideal to store the elements as a Vec.
When storing the database, it should be stored as an array.

Required Methods

fn size(&self) -> usize

Returns the size of the list (array.len(), vec.len(), array_list.size()).

fn get(&self, index: usize) -> T

Returns the value at the index (array[index], vec[index], array_list.get(index)).

Arguments

index - The index to get the value at.

fn set(&mut self, index: usize, val: T) -> Error<()>

Sets the value at the index.
array[index] = val, vec[index] = val, array_list.set(index, val).

Arguments

index - The index to get the value at.

val - The value to assign.

Returns

array : Ok.
vec : Ok if within size, otherwise Errors::OutOfBounds.
array_list: Ok if within size, otherwise Errors::OutOfBounds.