Module star_tracker_lib::util::list

A universal implementation of an array list, this is setup for a static ArrayList and Vec.

To enable the feature for Vec, you must compile with the feature “nix”.

Vec already is an implementation of a list, however, it is declared on the heap.
List provides a default implementation which can be used as a heap or stack object.
The main implementations are for ArrayList or Vec.
For embedded systems, it is advised to use ArrayList as it is declared on the stack.
For other systems, it is advised to use the vec implementation as it will never fill up.

Array List Structure

An array list is an oversized array with a pointer.
Like a stack, elements can be appended to the end and popped from the end quick and easily.
However, unlike a stack, an array_list is designed to access elements throughout like an array or linked list.
Consider the following diagram.

| 0 | 1 | 2 | 3 | 4 | 5 | pointer = 0
  x   x   x   x   x   x

Add an element (3).
| 0 | 1 | 2 | 3 | 4 | 5 | pointer = 1
  3   x   x   x   x   x
Add some elements (2, 5, 4).
| 0 | 1 | 2 | 3 | 4 | 5 | pointer = 4
  3   2   5   4   x   x

Returns 3: arraylist.get(0)
Returns 2: arraylist.get(1)
Returns 5: arraylist.get(2)
Returns 4: arraylist.get(3)

Memory Allocation

The heap can cause problems on microprocessors;

It takes time to allocate memory to the heap as oppose to the stack which is instant.

The size of the program is variable and it is hard to predict if the system will run out of memory.

Memory will be fragmented and scattered.
Consider running a program for a long time which allocates to a single memory location:

The program allocates every location.
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 |
  u   u   u   u   u   u   u   u   u   u   Free = 0

Some memory is free.
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 |
  u   u   _   u   u   _   u   _   u   u    Free = 3

Attempt to allocate an array of size 3.
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 |
  u   u   _   u   u   _   u   _   u   u    Free = 3
ERROR as the memory is fragmented and the array cannot fit.

Using a stack ensures that all the memory is available to the processor.
In big systems (i.e. computer), this is not nessisary to consider due to the processor speed and memory size.

Modules

array_list

Implementation of ArrayList.

list_iterator

Implementation of ListIterator.

vec

Addition to Linear Lookup for Vec.

Structs

ArrayList

An implementation of an Array List with no allocations to the heap.

ListIterator

Due to limitations with lifetimes and types, the default ToIterator trait cannot be used.
Instead the iterator statically creates a new iterator from an input list.

Traits

List

A set of instructions for a list/arraylist data structure.