Struct star_tracker_lib::util::list::ArrayList

pub struct ArrayList<T, const N: usize> { /* private fields */ }

An implementation of an Array List with no allocations to the heap.

Choose this data structure if:

You need a resizable, linear data structure of which you can access elements with an index.

You have limited memory on your device (i.e. microprocessor).

You need a reliable system which has a known memory usage.

You have a slow processor which will make heap allocations slow.

Example

The following example will show the versatility of an array_list.

use star_tracker_lib::util::list::ArrayList;
use star_tracker_lib::util::list::List;

// Creates an ArrayList which stores u32's, can store a maximum of 3 elements and is empty.
const max_capacity: usize = 3;
let mut lst: ArrayList<u32, max_capacity> = ArrayList::new();
assert!(lst.is_empty());

// Adds elements to the back of the list.
lst.push_back(2); // 2
lst.push_back(2); // 2, 2
lst.push_back(1); // 2, 2, 1

assert!(lst.is_full());
assert_eq!(lst.size(), 3);

// Assigns the value at the provided index.
lst.set(0, 3);    // 321

// Sorts the list with a custom sorting function.
fn sort_ascending ( left: & u32, right: & u32 ) -> bool	{  return left < right;  }
lst.sort_order(sort_ascending); // 1, 2, 3

// Verifies the list is in sorted order and removes each element.
assert_eq!(lst.pop_back(), Ok(3));  // 1, 2 | 3
assert_eq!(lst.pop_back(), Ok(2));  // 1    | 2
assert_eq!(lst.pop_back(), Ok(1));  //      | 1

// Slot fits elements into the array given the custom sorting function.
// As the core of this struct uses an array, consider this should not be used on a large array due to complexity and time.
lst.slot(10, sort_ascending); // 10
lst.slot(5, sort_ascending);  // 5, 10
lst.slot(1, sort_ascending);  // 1, 5, 10

// The list can also be created from an array.
// This is a useful function for test harnesses.
let mut lst_2: ArrayList<u32, 5> = ArrayList::from_array(&[1, 1, 2, 3, 5]);
 
// You can remove an element at an index and shuffle everything down.
// This is not recommended in a large list as shuffling everything takes time.
assert_eq!(lst.pop(0), Ok(1)); // 1, 2, 3, 5, _
 
// You can find elements which match between two lists of any type.
pub fn equal ( a: &u32, b: &u32 ) -> bool { return a == b; }
let lst_3 = ArrayList::from_array(&[2, 3, 5, 8]);
let mut lst_4: ArrayList<u32, 10> = ArrayList::new();
lst_2.find_match(&lst_3, equal, &mut lst_4);

assert_eq!(lst_4.size(), 3);
assert_eq!(lst_4.get(0), 2);
assert_eq!(lst_4.get(1), 3);
assert_eq!(lst_4.get(2), 5);

 
// You can also reduce your array to only have similarities to the other list.
lst_2.remove_diff(&lst_3, equal);
assert_eq!(lst_2.size(), 3);
assert_eq!(lst_2.get(0), 2);
assert_eq!(lst_2.get(1), 3);
assert_eq!(lst_2.get(2), 5);

// For extra memory security, you can reuse the same array by using clear().
lst_4.clear();
assert_eq!(lst_4.size(), 0);

// You can also clone the array (safely).
let lst_2_cloned = lst_2.clone();
assert_eq!(lst_2_cloned.size(), 3);
assert_eq!(lst_2_cloned.get(0), 2);
assert_eq!(lst_2_cloned.get(1), 3);
assert_eq!(lst_2_cloned.get(2), 5);

Implementations

impl<T, const N: usize> ArrayList<T, N>

pub fn new() -> ArrayList<T, N>

Constructor Use new or from_array to ensure the list is correctly setup.
This will create a new list with the given capacity and type with zero size.

Examples

use star_tracker_lib::util::list::ArrayList;

let list: ArrayList<u8, 10> = ArrayList::new(); // Creates a list of size 0, capacity 10.

impl<T, const N: usize> ArrayList<T, N>where T: Copy,

pub fn from_array(array: &[T; N]) -> Self

Creates an array list from an array.

Arguments

array - The array to copy.

Returns

Array < Capacity A copy of the array.
Array > Capacity Errors::InvalidSize

Examples

use star_tracker_lib::util::list::ArrayList;
use star_tracker_lib::util::list::List;

let array: [u32; 5] = [2, 4, 6, 8, 10];
let list: ArrayList<u32, 5> = ArrayList::from_array(&array);
assert_eq!(list.get(0), 2);
assert_eq!(list.get(1), 4);
assert_eq!(list.get(2), 6);
assert_eq!(list.get(3), 8);
assert_eq!(list.get(4), 10);