Trait star_tracker_lib::util::list::List

pub trait List<T> {

    // Required methods
    fn capacity(&self) -> usize;
    fn size(&self) -> usize;
    fn is_full(&self) -> bool;
    fn is_empty(&self) -> bool;
    fn get(&self, index: usize) -> T;
    fn set(&mut self, index: usize, value: T) -> Error<()>;
    fn push_back(&mut self, value: T) -> Error<()>;
    fn pop_back(&mut self) -> Error<T>;
    fn clear(&mut self);
    fn sort_order(&mut self, in_order: fn(_: &T, _: &T) -> bool);
    fn slot(&mut self, to_slot: T, in_order: fn(_: &T, _: &T) -> bool) -> bool;

    // Provided methods
    fn pop(&mut self, index: usize) -> Error<T> { ... }
    fn find_match(
        &self,
        list_b: &dyn List<T>,
        compare: fn(_: &T, _: &T) -> bool,
        out: &mut dyn List<T>
    ) -> Error<()> { ... }
    fn remove_diff(
        &mut self,
        other: &dyn List<T>,
        compare: fn(_: &T, _: &T) -> bool
    ) { ... }

}

A set of instructions for a list/arraylist data structure.

Vec already is an implementation of a list, however, it is declared on the heap.
List provides a default implementation which can be used as a heap or stack object.
The main implementations are for ArrayList or Vec.
For embedded systems, it is advised to use ArrayList as it is declared on the stack.
For other systems, it is advised to use the vec implementation as it will never fill up.

Memory Allocation

The heap can cause problems on microprocessors;

It takes time to allocate memory to the heap as oppose to the stack which is instant.

The size of the program is variable and it is hard to predict if the system will run out of memory.

Memory will be fragmented and scattered.
Consider running a program for a long time which allocates to a single memory location:

The program allocates every location.
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 |
  u   u   u   u   u   u   u   u   u   u   Free = 0

Some memory is free.
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 |
  u   u   _   u   u   _   u   _   u   u    Free = 3

Attempt to allocate an array of size 3.
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 |
  u   u   _   u   u   _   u   _   u   u    Free = 3
ERROR as the memory is fragmented and the array cannot fit.

Using a stack ensures that all the memory is available to the processor.
In big systems (i.e. computer), this is not nessisary to consider due to the processor speed and memory size.

Required Methods

fn capacity(&self) -> usize

Finds the max number of elements that can be stored in the list.
If invalid list, returns 0.

fn size(&self) -> usize

Finds how many elements are in the list.

Returns

The size.

fn is_full(&self) -> bool

Checks if the List is at maximum capacity.

Returns

True if full.

fn is_empty(&self) -> bool

Checks if the List is empty.

Returns

True if empty.

Examples

lst: ArrayList<UInt, 2> = ArrayList::new();

fn get(&self, index: usize) -> T

Gets the element at the specified index.

Arguments

‘index’ - The index of the element to receive.

Returns

The value at the index.

fn set(&mut self, index: usize, value: T) -> Error<()>

Sets the element at the specified index.

Arguments

‘index’ - The index of the element to receive.

‘value’ - The value to assign.

fn push_back(&mut self, value: T) -> Error<()>

Adds an element to the end of the list.

Arguments

‘value’ - the value to add to the end.

fn pop_back(&mut self) -> Error<T>

Removes an element from the end of the list.

Returns

The value removed.

fn clear(&mut self)

Sets the counter to 0 so all elements will be override and the list is essentially cleared.

fn sort_order(&mut self, in_order: fn(_: &T, _: &T) -> bool)

Sorts the list

Arguments

‘in_order’ - A function which returns TRUE if it is in order.

fn slot(&mut self, to_slot: T, in_order: fn(_: &T, _: &T) -> bool) -> bool

Slots an element into the list so it is in sorted order by shifting everything right.

Arguments

to_slot - The element to add.

in_order - The ordering method.

Returns

True if inserted, false if there is no space and it will trail the last element.

Provided Methods

fn pop(&mut self, index: usize) -> Error<T>

Removes element at index, reduces size of array list, moves everything left right of point.

Arguments

index - The index to remove.

Return

index < size The element at that point. size = 0 InvalidSize size =< index OutOfBounds

Examples

use star_tracker_lib::util::list::List;
use star_tracker_lib::util::list::ArrayList;
let mut list: ArrayList<usize, 4> = ArrayList::from_array(&[0, 1, 2, 3]);

assert_eq!((& mut list as &mut dyn List<usize>).pop(1), Ok(1));
assert_eq!(list.size(), 3);
assert_eq!(list.get(0), 0);
assert_eq!(list.get(1), 2);
assert_eq!(list.get(2), 3);

assert_eq!((& mut list as &mut dyn List<usize>).pop(0), Ok(0));
assert_eq!(list.size(), 2);
assert_eq!(list.get(0), 2);
assert_eq!(list.get(1), 3);

assert_eq!((&mut list as &mut dyn List<usize>).pop(1), Ok(3));
assert_eq!(list.size(), 1);
assert_eq!(list.get(0), 2);

assert!((&mut list as &mut dyn List<usize>).pop(1).is_err());
assert_eq!(list.size(), 1);
assert_eq!(list.get(0), 2);

assert_eq!((&mut list as &mut dyn List<usize>).pop(0), Ok(2));
assert_eq!(list.size(), 0);

assert!((&mut list as &mut dyn List<usize>).pop(0).is_err());
assert_eq!(list.size(), 0);

fn find_match( &self, list_b: &dyn List<T>, compare: fn(_: &T, _: &T) -> bool, out: &mut dyn List<T> ) -> Error<()>

Finds any associated items and appends self to out.

Arguments

list_b - A list to compare with list_a.

compare - A function to compare the two (usually a == b).

out - The output array which holds the similarities.

Returns

Errors::InvalidSize if the list is full and more elements should be added. Ok(()) otherwise.

Examples

use star_tracker_lib::util::list::List;
use star_tracker_lib::util::list::ArrayList;
let mut list_a: ArrayList<usize, 10> = ArrayList::new();
let list_b: ArrayList<usize, 3> = ArrayList::from_array(&[0, 1, 2]);
list_a.push_back(1);
list_a.push_back(2);
list_a.push_back(3);
fn func ( x: &usize, y: &usize ) -> bool {x == y};
let mut out: ArrayList<usize, 10> = ArrayList::new();

list_a.find_match ( &list_b, func, &mut out );
assert_eq!(out.size(), 2);
assert_eq!(out.get(0), 1);
assert_eq!(out.get(1), 2);

fn remove_diff( &mut self, other: &dyn List<T>, compare: fn(_: &T, _: &T) -> bool )

Removes all elements which do not share the similarities in the compare function and list.

Arguments

other - The list to compare to.

compare - The function to compare the lists together.

Examples

use star_tracker_lib::util::list::List;
use star_tracker_lib::util::list::ArrayList;
let mut list_a: ArrayList<usize, 10> = ArrayList::new();
let list_b: ArrayList::<usize, 3> = ArrayList::from_array(&[0, 1, 2]);
list_a.push_back(1);
list_a.push_back(2);
list_a.push_back(3);
fn func ( x: &usize, y: &usize ) -> bool {x == y};
let mut out: ArrayList<usize, 0> = ArrayList::new();

list_a.remove_diff( &list_b, func );
assert_eq!(out.size(), 0);