Module star_tracker_lib::util::maths

A wrapper for libm.
This is needed as libm is not included by default into the no_std package.

Traits

Maths

Functions

acos

Arccosine (f64)

acosf

Arccosine (f32)

asin

Arcsine (f64)

asinf

Arcsine (f32)

atan2

Arctangent of y/x (f64)

atan2f

Arctangent of y/x (f32)

ceil

Ceil (f64)

ceilf

Ceil (f32)

copysign

Sign of Y, magnitude of X (f64)

copysignf

Sign of Y, magnitude of X (f32)

cos

cosf

fabs

Absolute value (magnitude) (f64) Calculates the absolute value (magnitude) of the argument x, by direct manipulation of the bit representation of x.

fabsf

Absolute value (magnitude) (f32) Calculates the absolute value (magnitude) of the argument x, by direct manipulation of the bit representation of x.

floor

Floor (f64)

floorf

Floor (f32)

pow

powf

round

roundf

sin

sinf

sqrt

sqrtf