Trait star_tracker_lib::util::test::TestEqual

pub trait TestEqual {
    // Required method
    fn test_close(&self, other: &Self, precision: Decimal) -> bool;

    // Provided methods
    fn test_equal(&self, other: &Self) -> bool
       where Self: Debug { ... }
    fn assert_close(&self, other: &Self, precision: Decimal)
       where Self: Debug { ... }
}

An equality test with a precision variable so the equality can be better checked.

Required Methods

fn test_close(&self, other: &Self, precision: Decimal) -> bool

An equality test with a precision variable so the equality can be better checked.

Arguments

other - The element to compare with.

precision - The tolerance.

Returns

True if equal.

Provided Methods

fn test_equal(&self, other: &Self) -> boolwhere Self: Debug,

Auto implemented function which calls test_equal with precision of DECIMAL_PRECISION_TEST.

Arguments

other - The element to compare to.

Returns

True if equal.

fn assert_close(&self, other: &Self, precision: Decimal)where Self: Debug,

This automatically calls an assert if test_close fails.
The call stack will display the line which calls this.