Module star_tracker_lib::util::units

This is a set of units and coordinates which are used throught the project.

Try to use these as much as possible as it is more explicit and avoids silly mistakes.

Modules

angle_axis

Implementation of AngleAxis.

angles

Implementation of Degrees, Radians and Hours.

bit_field

Implementation of BitField.

crp

Implementation of CRP.

equatorial

Implementation of Equatorial.

matrix

Implementation of Matrix.

quaternion

Implementation of Quaternion.

unit_ops

Implementation of Mul, Add, Sub, Div and Display

vector2

Implementation for Vector2.

vector3

Implementation for Vector3.

Structs

AngleAxis

An axis describing the pivot point and an angle specifying how much to rotate around.

BitField

Stores a bit field and can do bit operations easily. Example

CRP

A CRP is an old method of describing a rotation (poorly).

Degrees

An angle in degrees.

Equatorial

Equatorial Coordinates are coordinates which define a point on a unit sphere.

Hours

An angle in 24 hours.

MatPos

The coordinates for the matrix.

Match

A predicted input, possible match output and a weighting of uncertaincy.

Matrix

An n x m matrix declared on the stack.
Use for an extensive set of matrix equations and modifications.

Quaternion

Represents a 3D rotation without singularity.

Radians

An angle in radians.

Vector2

2D Cartesian Coordinates in Decimal format.

Vector2Int

An integer version of Vector2.

Vector3

A 3d cartesian vector/point.

Enums

BitCompare

Specifies how to compare bit fields.
Refer to BitField for uses.

Type Definitions

Pixel

Alias for Vector2Int (x: uusize, y: usize).