Struct star_tracker_lib::util::units::BitField

pub struct BitField(pub UInt);

Stores a bit field and can do bit operations easily. Example

use star_tracker_lib::util::units::BitField;
use star_tracker_lib::util::units::BitCompare;

let mut field_lsb : BitField = BitField(0b0011); 
let field_msb : BitField = BitField(0b1100); 
let field_on  : BitField = BitField(0b1111); 
let field_off : BitField = BitField(0b0000);

// Any ensures at least one bit is the same:
assert!(field_lsb.compare(BitCompare::Any(field_on)));   // 0011 & 1111 = true 
assert!(!field_lsb.compare(BitCompare::Any(field_off))); // 0011 & 0000 = false
assert!(!field_lsb.compare(BitCompare::Any(field_msb))); // 0011 & 1100 = false

// All ensures all bits are equal:
assert!(!field_lsb.compare(BitCompare::All(field_on)));   // 0011 == 1111 = false
assert!(field_lsb.compare(BitCompare::All(field_lsb)));  // 0011 == 0011 = true

// Bits can easily be set:
field_lsb.set(3, true); // 0111.
field_lsb.set(4, true); // 1111.
assert!(!field_lsb.compare(BitCompare::All(field_on)));   // 1111 == 1111 = true

Tuple Fields
0: UInt
Implementations

impl BitField

pub const FIELDS: usize = 4usize

The number of elements the bit field can represent.

pub const ALL: BitField = _

pub fn compare(self, field: BitCompare) -> bool

Checks if there are any similarities between the bitfields.
Returns true if the a bit is true in both.

pub fn set(&mut self, index: usize, value: bool)

Sets the bit to the desired value.