Struct star_tracker_lib::util::units::Equatorial

pub struct Equatorial {
    pub ra: Radians,
    pub dec: Radians,
}

Equatorial Coordinates are coordinates which define a point on a unit sphere.

Right Ascension (ra) is defined as the angle around the equator from 0 to 2PI (Don’t use 0h to 24h or 0 to 360 deg).
Declination (dec) is defined as the angle from -PI to PI (Don’t use -90 to 90 deg).
Equatorial must be converted to cartesian Vector3 to perform any equations.
Use Equatorial as an efficient way of storing positions as it is 2 numbers instead of 3.

Fields
ra: Radians

Right Ascention (along the equator [0, 360)).

dec: Radians

Declination (from pole to pole [-90, 90]).

Implementations

impl Equatorial

pub const fn zero() -> Equatorial

Returns an equatorial coordinate at ra: 0, dec: 0 (on the equator at 0 hours).
Useful for initialization or testing.

Example

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;

assert_eq!(Equatorial{ra: Radians(0.0), dec: Radians(0.0)}, Equatorial::zero());

pub const fn north() -> Equatorial

Returns an equatorial coordinate at ra: 0, dec: 90 (north). Useful for initialization and testing.

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::units::Degrees;

assert_eq!(Equatorial{ra: Radians(0.0), dec: Degrees(90.0).as_radians()}, Equatorial::north());

pub const fn south() -> Equatorial

Returns an equatorial coordinate at ra: 0, dec: -90 (south). Useful for initialization and testing.

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::units::Degrees;

assert_eq!(Equatorial{ra: Radians(0.0), dec: Degrees(-90.0).as_radians()}, Equatorial::south());

pub fn range_dec() -> RangeInclusive<Radians>

The range for declination.
Useful for random generation testing.

pub fn range_ra() -> RangeInclusive<Radians>

The range for right ascention Useful for random genration testing.

pub fn get_phi(&self) -> Radians

Returns dec + 180 degrees to the declination to match with spherical coordinates.

Example

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::aliases::M_PI;

let equ = Equatorial { ra: Radians(0.0), dec: Radians(-M_PI / 2.0) };
assert_eq!( equ.get_phi(), Radians(0.0) );

pub fn set_phi(&mut self, angle: Radians)

Sets phi (start at z = +1).

Example

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::aliases::M_PI;

let mut equ = Equatorial { ra: Radians(0.0), dec: Radians(0.0) };
equ.set_phi( Radians(M_PI / 4.0) );
assert_eq!( equ.dec, Radians(-M_PI / 4.0) );

pub fn angle_distance(&self, other: Equatorial) -> Radians

USE Vector3.angular_distance IF YOU HAVE A CARTESIAN, CONVERTING TO EQUATORIAL HAS A SINGULARITY!!!
Finds the angle between the 2 points on a sphere.
Just a shortcut to Vector3 equivalent.

Example

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::aliases::M_PI;

let mut equ1 = Equatorial { ra: Radians(M_PI), dec: Radians(M_PI / 2.0) };
let mut equ2 = Equatorial { ra: Radians(-M_PI), dec: Radians(-M_PI / 2.0) }; // 180 degrees because of dec
assert_eq!(equ1.angle_distance(equ2), Radians(M_PI));

equ1 = Equatorial { ra: Radians(0.0), dec: Radians(M_PI / 4.0) };
equ2 = Equatorial { ra: Radians(M_PI), dec: Radians(M_PI / 4.0) };
assert_eq!(equ1.angle_distance(equ2), Radians(M_PI/2.0));

pub fn planar_distance(&self, other: Equatorial) -> Decimal

Finds the distance between the 2 points cutting a strait 2d line through the sphere.

Example

use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::aliases::M_PI;

let mut equ1 = Equatorial { ra: Radians(M_PI), dec: Radians(M_PI / 2.0) };
let mut equ2 = Equatorial { ra: Radians(-M_PI), dec: Radians(-M_PI / 2.0) }; // 180 degrees because of dec
assert_eq!(equ1.planar_distance(equ2), 2.0);

equ1 = Equatorial { ra: Radians(0.0), dec: Radians(M_PI / 4.0) };
equ2 = Equatorial { ra: Radians(M_PI), dec: Radians(M_PI / 4.0) };
assert_eq!(equ1.planar_distance(equ2), 1.4142135623730951);

pub fn to_vector3(&self) -> Vector3

Gets a 3d cartesian coordinate from the equatorial unit sphere.

Example

use star_tracker_lib::util::units::Equatorial;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::aliases::M_PI;
let equ = Equatorial { ra: Radians(M_PI / 4.0), dec: Radians(M_PI / 4.0) };
let cart = equ.to_vector3();
assert!((cart.x - 0.5).abs() < 0.0001);
assert!((cart.y - 0.5).abs() < 0.0001);
assert!((cart.z - 0.7071067812).abs() < 0.0001);