Struct star_tracker_lib::util::units::Matrix

pub struct Matrix<const ROW: usize, const COLUMN: usize> { /* private fields */ }

An n x m matrix declared on the stack.
Use for an extensive set of matrix equations and modifications.

Some Uses

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;
use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::aliases::Decimal;

let mut matrix: Matrix<3, 2> = Matrix::new(); // A 0 matrix 3 high (3 rows), 2 wide (2 columns).
assert_eq!(matrix.width(), 2);
assert_eq!(matrix.height(), 3);

let eye:    Matrix<2, 2> = Matrix::identity(); // An identity 2x2 matrix.
assert_eq!(eye.get(MatPos{row: 0, col: 0}), 1.0);
assert_eq!(eye.get(MatPos{row: 0, col: 1}), 0.0);
assert_eq!(eye.get(MatPos{row: 1, col: 0}), 0.0);
assert_eq!(eye.get(MatPos{row: 1, col: 1}), 1.0);
assert_eq!(eye.trace(), 2.0); // You can do trace of any size. 
assert_eq!(eye.determinate(), 1.0); // You can do determinate of up to and including 4x4. 

matrix.set(MatPos{row: 1, col: 0}, 10.0); // sets row 1, col 0 to 10.
let mut trans: Matrix<2, 3> = matrix.transposed(); // You can transpose.

trans.insert(MatPos{row: 0, col: 1}, &eye); // Insert the identity matrix into this one.
// Trans (after insert):
// 1	10	0
// 0	1	0

// Matrix multiplication can be done for: 3x3, 3x4, 4x4
// There is also adjoint.
 

// You can create matrices from single values and vectors:
let small_mat: Matrix<1,1> = Matrix::from_decimal(10.0);
let decimal  : Decimal     = small_mat.to_decimal();     // 10.0.

let vec_mut  : Matrix<3,1> = Vector3::to_matrix_column(&Vector3{x: 10.0, y: 20.0, z: 30.0}); // row vec.
let vector   : Vector3     = vec_mut.to_vector3();

let vec_mut_homo : Matrix<4,1> = Vector3::to_matrix_column_homo(&Vector3{x: 10.0, y: 20.0, z: 30.0}); // homogenous column matrix.
let vector       : Vector3     = vec_mut_homo.to_vector3(); // divides by the 4th value.

 

Column is the width or x.
Row is the height or y.

Implementations

impl<const ROW: usize, const COLUMN: usize> Matrix<ROW, COLUMN>

pub fn new() -> Matrix<ROW, COLUMN>

Default Constructor. Initializes matrix to 0.

Example

use star_tracker_lib::util::units::Matrix;

let mat3x4: Matrix<3, 4> = Matrix::new(); // Matrix with 3 rows and 4 columns

impl<const S: usize> Matrix<S, S>

pub fn identity() -> Matrix<S, S>

Contructs an identity matrix.

impl<const ROW: usize, const COLUMN: usize> Matrix<ROW, COLUMN>

pub fn width(&self) -> usize

Returns the width of the matrix.

Example

use star_tracker_lib::util::units::Matrix;

let mat3x4: Matrix<3, 4> = Matrix::new();
assert_eq!(mat3x4.width(), 4);

pub fn height(&self) -> usize

Returns the width of the matrix.

Example

use star_tracker_lib::util::units::Matrix;

let mat3x4: Matrix<3, 4> = Matrix::new();
assert_eq!(mat3x4.height(), 3);

pub fn get(&self, pos: MatPos) -> Decimal

Returns the value at the given index.

Example

use star_tracker_lib::util::err::Errors;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat3x4: Matrix<3, 4> = Matrix::new();
mat3x4.set(MatPos{row: 1, col: 2}, 3.12);
mat3x4.set(MatPos{row: 2, col: 3}, 5.23);
assert_eq!(mat3x4.get(MatPos{row: 1, col: 2}), 3.12);
assert_eq!(mat3x4.get(MatPos{row: 2, col: 3}), 5.23);
assert_eq!(mat3x4.get(MatPos{row: 0, col: 0}), 0.0);

pub fn set(&mut self, pos: MatPos, value: Decimal)

Returns the value at the given index.

Example

use star_tracker_lib::util::err::Errors;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat3x4: Matrix<3, 4> = Matrix::new();
mat3x4.set(MatPos{row: 1, col: 2}, 3.12);
mat3x4.set(MatPos{row: 2, col: 3}, 5.23);
assert_eq!(mat3x4.get(MatPos{row: 1, col: 2}), 3.12);
assert_eq!(mat3x4.get(MatPos{row: 2, col: 3}), 5.23);
assert_eq!(mat3x4.get(MatPos{row: 0, col: 0}), 0.0);

pub fn transposed(&self) -> Matrix<COLUMN, ROW>

Flips the rows and columns.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat3x4: Matrix<3,4> = Matrix::new();
mat3x4.set(MatPos{row: 0, col: 0}, 0.0);
mat3x4.set(MatPos{row: 0, col: 1}, 0.1);
mat3x4.set(MatPos{row: 1, col: 0}, 1.0);
mat3x4.set(MatPos{row: 1, col: 1}, 1.1);
mat3x4.set(MatPos{row: 2, col: 3}, 2.3);
let mat4x3 = mat3x4.transposed();
assert_eq!(mat4x3.width(), 3);
assert_eq!(mat4x3.height(), 4);
assert_eq!(mat3x4.get(MatPos{row:0, col:0}), mat4x3.get(MatPos{row:0, col:0}));
assert_eq!(mat3x4.get(MatPos{row:0, col:1}), mat4x3.get(MatPos{row:1, col:0}));
assert_eq!(mat3x4.get(MatPos{row:1, col:0}), mat4x3.get(MatPos{row:0, col:1}));
assert_eq!(mat3x4.get(MatPos{row:1, col:1}), mat4x3.get(MatPos{row:1, col:1}));
assert_eq!(mat3x4.get(MatPos{row:2, col:3}), mat4x3.get(MatPos{row:3, col:2}));

pub fn insert<const C_2: usize, const R_2: usize>( &mut self, pos: MatPos, other: &Matrix<C_2, R_2> ) -> Error<()>

Inserts the matrix into the specified location on this matrix.

Returns

Ok(()) if the matrix fits.
Err(Errors::InvalidSize) if the matrix does not fit at the location specified.
On failure, the matrix is not inserted.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat4x4: Matrix<4,4> = Matrix::new();
let mut insert: Matrix<3,2> = Matrix::new();

insert.set(MatPos{row: 0, col: 0}, 1.0);
insert.set(MatPos{row: 0, col: 1}, 2.0);
insert.set(MatPos{row: 1, col: 0}, 3.0);
insert.set(MatPos{row: 1, col: 1}, 4.0);
insert.set(MatPos{row: 2, col: 0}, 5.0);
insert.set(MatPos{row: 2, col: 1}, 6.0);

mat4x4.insert(MatPos{row: 1, col: 1}, &insert);

assert_eq!(mat4x4.get(MatPos{row: 0, col: 0}), 0.0); // first element is 1,1
assert_eq!(mat4x4.get(MatPos{row: 1, col: 1}), 1.0);
assert_eq!(mat4x4.get(MatPos{row: 1, col: 2}), 2.0);
assert_eq!(mat4x4.get(MatPos{row: 2, col: 1}), 3.0);
assert_eq!(mat4x4.get(MatPos{row: 2, col: 2}), 4.0);
assert_eq!(mat4x4.get(MatPos{row: 3, col: 1}), 5.0);
assert_eq!(mat4x4.get(MatPos{row: 3, col: 2}), 6.0);

impl<const S: usize> Matrix<S, S>

pub fn trace(&self) -> Decimal

Finds the trace of the matrix (sum of diagonal). Must be a square matrix.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat4x4: Matrix<4,4> = Matrix::new();
mat4x4.set(MatPos{row: 0, col: 0}, 1.0);
mat4x4.set(MatPos{row: 1, col: 1}, 2.0);
mat4x4.set(MatPos{row: 2, col: 2}, 4.0);
mat4x4.set(MatPos{row: 3, col: 3}, 8.0);

assert_eq!(mat4x4.trace(), 1.0 + 2.0 + 4.0 + 8.0);

impl Matrix<1, 1>

pub fn from_decimal(val: Decimal) -> Matrix<1, 1>

Converts the current value to a matrix.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mat1x1: Matrix<1,1> = Matrix::from_decimal(100.0);
assert_eq!(mat1x1.get(MatPos{row: 0, col: 0}), 100.0);

pub fn to_decimal(&self) -> Decimal

Converts the current value to a decimal.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat1x1: Matrix<1,1> = Matrix::new();
mat1x1.set(MatPos{row: 0, col: 0}, 100.0);
assert_eq!(mat1x1.to_decimal(), 100.0);

pub fn determinate(&self) -> Decimal

Returns the single element.

Example

use star_tracker_lib::util::units::Matrix;

use star_tracker_lib::util::units::MatPos;
let mut mat1x1: Matrix<1,1> = Matrix::new();
mat1x1.set(MatPos{row: 0, col: 0}, 10.0);
assert_eq!(mat1x1.determinate(), 10.0);

impl Matrix<2, 2>

pub fn determinate(&self) -> Decimal

Recursively finds the determinate.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat2x2: Matrix<2,2> = Matrix::new();
mat2x2.set(MatPos{row: 0, col: 0}, 2.0);
mat2x2.set(MatPos{row: 1, col: 0}, 3.0);
mat2x2.set(MatPos{row: 0, col: 1}, 4.0);
mat2x2.set(MatPos{row: 1, col: 1}, 5.0);
assert_eq!(mat2x2.determinate(), 2.0 * 5.0 - 3.0 * 4.0);
assert_eq!(mat2x2.determinate(), -2.0);

impl Matrix<3, 3>

pub fn determinate(&self) -> Decimal

Recursively finds the determinate.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat3x3: Matrix<3,3> = Matrix::new();
mat3x3.set(MatPos{row: 0, col: 0}, 5.0);
mat3x3.set(MatPos{row: 0, col: 1}, 4.0);
mat3x3.set(MatPos{row: 0, col: 2}, 3.0);
mat3x3.set(MatPos{row: 1, col: 0}, 8.0);
mat3x3.set(MatPos{row: 1, col: 1}, 9.0);
mat3x3.set(MatPos{row: 1, col: 2}, 10.0);
mat3x3.set(MatPos{row: 2, col: 0}, 100.0);
mat3x3.set(MatPos{row: 2, col: 1}, 2.0);
mat3x3.set(MatPos{row: 2, col: 2}, 6.0);
assert_eq!(mat3x3.determinate(), 1326.0);

pub fn adjoint(&self) -> Matrix<3, 3>

Finds the adjoint matrix. The adjoint matrix replaces each cell with the determinate from looking at that cell.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat3x3: Matrix<3,3> = Matrix::new();
mat3x3.set(MatPos{row: 0, col: 0}, 5.0);
mat3x3.set(MatPos{row: 0, col: 1}, 4.0);
mat3x3.set(MatPos{row: 0, col: 2}, 3.0);
mat3x3.set(MatPos{row: 1, col: 0}, 8.0);
mat3x3.set(MatPos{row: 1, col: 1}, 9.0);
mat3x3.set(MatPos{row: 1, col: 2}, 10.0);
mat3x3.set(MatPos{row: 2, col: 0}, 100.0);
mat3x3.set(MatPos{row: 2, col: 1}, 2.0);
mat3x3.set(MatPos{row: 2, col: 2}, 6.0);

let mut res: Matrix<3,3> = Matrix::new();
res.set(MatPos{row: 0, col: 0}, 34.0);
res.set(MatPos{row: 0, col: 1}, 952.0);
res.set(MatPos{row: 0, col: 2}, -884.0);
res.set(MatPos{row: 1, col: 0}, -18.0);
res.set(MatPos{row: 1, col: 1}, -270.0);
res.set(MatPos{row: 1, col: 2}, 390.0);
res.set(MatPos{row: 2, col: 0}, 13.0);
res.set(MatPos{row: 2, col: 1}, -26.0);
res.set(MatPos{row: 2, col: 2}, 13.0);

assert_eq!(res, mat3x3.adjoint());

impl Matrix<4, 4>

pub fn determinate(&self) -> Decimal

Recursively finds the determinate.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let mut mat4x4: Matrix<4,4> = Matrix::new();

mat4x4.set(MatPos{row: 0, col: 0}, 5.0);
mat4x4.set(MatPos{row: 0, col: 1}, 10.0);
mat4x4.set(MatPos{row: 0, col: 2}, 50.0);
mat4x4.set(MatPos{row: 0, col: 3}, 2.0);

mat4x4.set(MatPos{row: 1, col: 0}, 8.0);
mat4x4.set(MatPos{row: 1, col: 1}, 20.0);
mat4x4.set(MatPos{row: 1, col: 2}, 60.0);
mat4x4.set(MatPos{row: 1, col: 3}, 11.0);

mat4x4.set(MatPos{row: 2, col: 0}, 100.0);
mat4x4.set(MatPos{row: 2, col: 1}, 30.0);
mat4x4.set(MatPos{row: 2, col: 2}, 70.0);
mat4x4.set(MatPos{row: 2, col: 3}, 103.0);

mat4x4.set(MatPos{row: 3, col: 0}, 50.0);
mat4x4.set(MatPos{row: 3, col: 1}, 40.0);
mat4x4.set(MatPos{row: 3, col: 2}, 80.0);
mat4x4.set(MatPos{row: 3, col: 3}, 53.0);

assert_eq!(mat4x4.determinate(), 340800.0);

impl Matrix<3, 1>

pub fn to_vector3(&self) -> Vector3

Converts the matrix into a Vector3.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;
use star_tracker_lib::util::units::Vector3;

let mut mat3x1: Matrix<3,1> = Matrix::new();
mat3x1.set(MatPos{row: 0, col: 0}, 1.0);
mat3x1.set(MatPos{row: 1, col: 0}, 2.0);
mat3x1.set(MatPos{row: 2, col: 0}, 3.0);
assert_eq!(mat3x1.to_vector3(), Vector3{x: 1.0, y: 2.0, z: 3.0});

impl Matrix<4, 1>

pub fn to_vector3(&self) -> Vector3

Converts the matrix into a Vector3.

Example

use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;
use star_tracker_lib::util::units::Vector3;

let mut mat4x1: Matrix<4,1> = Matrix::new();
mat4x1.set(MatPos{row: 0, col: 0}, 1.0);
mat4x1.set(MatPos{row: 1, col: 0}, 2.0);
mat4x1.set(MatPos{row: 2, col: 0}, 3.0);
mat4x1.set(MatPos{row: 3, col: 0}, 0.5);
assert_eq!(mat4x1.to_vector3(), Vector3{x: 2.0, y: 4.0, z: 6.0});

impl Matrix<3, 3>

pub fn multiply(&self, rhs: Vector3) -> Vector3

Multiplies this matrix by the provided vector.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let pt = Vector3{x: 1.0, y: 2.0, z: 3.0};
let mut mat3x3: Matrix<3,3> = Matrix::new();

// x is the addition of all.
mat3x3.set(MatPos{row: 0, col: 0}, 1.0);
mat3x3.set(MatPos{row: 0, col: 1}, 1.0);
mat3x3.set(MatPos{row: 0, col: 2}, 1.0);

// y is 2y + z.
mat3x3.set(MatPos{row: 1, col: 0}, 0.0);
mat3x3.set(MatPos{row: 1, col: 1}, 2.0);
mat3x3.set(MatPos{row: 1, col: 2}, 1.0);

// z is 0.5x + 3.0y + 2.0z.
mat3x3.set(MatPos{row: 2, col: 0}, 0.5);
mat3x3.set(MatPos{row: 2, col: 1}, 3.0);
mat3x3.set(MatPos{row: 2, col: 2}, 2.0);

let expected = Vector3
{
	x: pt.x + pt.y + pt.z,
	y: 2.0 * pt.y + pt.z,
	z: 0.5 * pt.x + 3.0 * pt.y + 2.0 * pt.z,
};
assert_eq!(mat3x3.multiply(pt), expected);

impl Matrix<3, 4>

pub fn multiply(&self, rhs: Vector3) -> Vector3

Multiplies this matrix by the provided vector.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let pt = Vector3{x: 1.0, y: 2.0, z: 3.0};
let mut mat4x4: Matrix<3,4> = Matrix::new();

// x is the addition of all + 1.
mat4x4.set(MatPos{row: 0, col: 0}, 1.0);
mat4x4.set(MatPos{row: 0, col: 1}, 1.0);
mat4x4.set(MatPos{row: 0, col: 2}, 1.0);
mat4x4.set(MatPos{row: 0, col: 3}, 1.0);

// y is 2y + z.
mat4x4.set(MatPos{row: 1, col: 0}, 0.0);
mat4x4.set(MatPos{row: 1, col: 1}, 2.0);
mat4x4.set(MatPos{row: 1, col: 2}, 1.0);
mat4x4.set(MatPos{row: 1, col: 3}, 0.0);

// z is 0.5x + 3.0y + 2.0z + 1.
mat4x4.set(MatPos{row: 2, col: 0}, 0.5);
mat4x4.set(MatPos{row: 2, col: 1}, 3.0);
mat4x4.set(MatPos{row: 2, col: 2}, 2.0);
mat4x4.set(MatPos{row: 2, col: 3}, 1.0);

let expected = Vector3
{
	x: pt.x + pt.y + pt.z + 1.0,
	y: 2.0 * pt.y + pt.z,
	z: 0.5 * pt.x + 3.0 * pt.y + 2.0 * pt.z + 1.0,
};
assert_eq!(mat4x4.multiply(pt), expected);

impl Matrix<4, 4>

pub fn multiply(&self, rhs: Vector3) -> Vector3

Multiplies this matrix by the provided vector.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let pt = Vector3{x: 1.0, y: 2.0, z: 3.0};
let mut mat3x4: Matrix<4,4> = Matrix::new();

// x is the addition of all.
mat3x4.set(MatPos{row: 0, col: 0}, 1.0);
mat3x4.set(MatPos{row: 0, col: 1}, 1.0);
mat3x4.set(MatPos{row: 0, col: 2}, 1.0);
mat3x4.set(MatPos{row: 0, col: 3}, 0.0);

// y is 2y + z + 1.
mat3x4.set(MatPos{row: 1, col: 0}, 0.0);
mat3x4.set(MatPos{row: 1, col: 1}, 2.0);
mat3x4.set(MatPos{row: 1, col: 2}, 1.0);
mat3x4.set(MatPos{row: 1, col: 3}, 1.0);

// z is 0.5x + 3.0y + 2.0z.
mat3x4.set(MatPos{row: 2, col: 0}, 0.5);
mat3x4.set(MatPos{row: 2, col: 1}, 3.0);
mat3x4.set(MatPos{row: 2, col: 2}, 2.0);
mat3x4.set(MatPos{row: 2, col: 3}, 0.0);

// w is 2x + 2y + 2z + 2.
mat3x4.set(MatPos{row: 3, col: 0}, 2.0);
mat3x4.set(MatPos{row: 3, col: 1}, 2.0);
mat3x4.set(MatPos{row: 3, col: 2}, 2.0);
mat3x4.set(MatPos{row: 3, col: 3}, 2.0);

let w = 2.0 * (pt.x + pt.y + pt.z) + 2.0;
let expected = Vector3
{
	x: pt.x + pt.y + pt.z,
	y: 2.0 * pt.y + pt.z + 1.0,
	z: 0.5 * pt.x + 3.0 * pt.y + 2.0 * pt.z,
} / w;
assert_eq!(mat3x4.multiply(pt), expected);