Struct star_tracker_lib::util::units::Quaternion

pub struct Quaternion {
    pub w: Decimal,
    pub x: Decimal,
    pub y: Decimal,
    pub z: Decimal,
}

Represents a 3D rotation without singularity.

This method uses the LEFT HAND COORDINATE SYSTEM.
This is the most ideal and confusing coordinate system.

Fields
w: Decimalx: Decimaly: Decimalz: Decimal
Implementations

impl Quaternion

pub fn rotate_point(self, pt: Vector3) -> Vector3

Rotates a point around this quaternion.

Example

use star_tracker_lib::util::units::Quaternion;
use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::units::AngleAxis;
use star_tracker_lib::util::aliases::M_PI;

// Rotates point around the z axis where the point is 10 degrees x of z
let axis = Vector3{x: 0.0, y: 0.0, z: 1.0};
let angle = Radians(M_PI / 4.0); // 45 degrees
let angle_axis = AngleAxis{angle: angle, axis: axis};
let offset_angle = Radians(M_PI / 18.0); // 10 degrees x of z.
let pt = Vector3{x: (offset_angle).sin(), y: 0.0, z: (offset_angle).cos()};
let q = angle_axis.to_quaternion();
let rotated = q.rotate_point(pt);
let expected = Vector3{
						x: (offset_angle).sin() * (Radians(M_PI/2.0) + angle).sin(),
						y: (offset_angle).sin() * (Radians(M_PI/2.0) + angle).cos(),
						z: pt.z};
assert!((rotated.x - expected.x).abs() < 0.0001);
assert!((rotated.y - expected.y).abs() < 0.0001);
assert!((rotated.z - expected.z).abs() < 0.0001);

pub fn to_angle_axis(&self) -> AngleAxis

Creates an angle axis from this quaternion.

Example

use star_tracker_lib::util::units::Quaternion;
use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::units::AngleAxis;
use star_tracker_lib::util::aliases::M_PI;

let angle = Radians(M_PI / 2.0);
let axis = Vector3{x: 1.0, y: 0.0, z: 0.0};
let angle_axis = AngleAxis{angle: angle, axis: axis};
let q = angle_axis.to_quaternion();
assert_eq!(angle_axis, q.to_angle_axis());

pub fn conjugate(&self) -> Quaternion

Finds the conjugate of this quaternion.

use star_tracker_lib::util::units::Quaternion;
let q = Quaternion{w: 1.0, x: 2.0, y: 3.0, z: 4.0};
assert_eq!(q.conjugate(), Quaternion{w: 1.0, x: -2.0, y: -3.0, z: -4.0});

pub fn dot(&self, other: Quaternion) -> Decimal

Finds the dot product of the quaternions.
This is a useful way of comparing 2 rotations without worrying about a singularity.
If the values are close, the rotated vector will be in the same location.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Quaternion;
use star_tracker_lib::util::units::AngleAxis;
use star_tracker_lib::util::units::Radians;

//Compares 2 quaternions which are identical but are represented incorrectly as angle axis.
let a_1 = AngleAxis{angle: Radians(90.000001), axis: Vector3{x: 1.0, y: 0.0, z: 0.0}};
let a_2 = AngleAxis{angle: Radians(-90.0001), axis: Vector3{x: -1.0, y: 0.0, z: 0.0}};
let a_3 = AngleAxis{angle: Radians(-90.0001), axis: Vector3{x: 0.0, y: 1.0, z: 0.0}};

let q_1 = a_1.to_quaternion();
let q_2 = a_2.to_quaternion();
let q_3 = a_3.to_quaternion();

assert!(1.0 - q_1.dot(q_2).abs() < 0.00001);
assert!(0.00001 < 1.0 - q_1.dot(q_3).abs());