Struct star_tracker_lib::util::units::Radians

pub struct Radians(pub Decimal);

An angle in radians.

Radians are the default measurement of angles.
Degrees and Hours are only used as a user interface.

use star_tracker_lib::util::units::Radians;
use star_tracker_lib::util::test::TestEqual;
use star_tracker_lib::util::aliases::DECIMAL_PRECISION;

let angle = Radians(1.2345);
angle.to_degrees();                                  // Easy to read format.
angle.to_hours();                                    // Standard Equatorial format.
 
angle.sin();                                         // Easier than angle.0.sin().
angle.cos();                                         // Easier than angle.0.cos().

assert_eq!(angle + Radians(1.0), Radians(2.2345));   // Can add standard decimal.
assert_eq!(angle - Radians(1.0), Radians(0.2345));   // Can multiply radians.

assert_eq!(angle * 2.0,          Radians(2.469));    // Can multiply standard decimal.
assert_eq!(angle * Radians(2.0), Radians(2.469));    // Can multiply radians.

assert_eq!(angle / 2.0,          Radians(0.61725));  // Can multiply standard decimal.
assert_eq!(angle / Radians(2.0), Radians(0.61725));  // Can multiply radians.

assert_eq!(-angle,               Radians(-1.2345));  // Can flip the sign.
assert_eq!((-angle).abs(),       1.2345);            // Can perform abs function.

assert_eq!(angle, angle + Radians(DECIMAL_PRECISION / 2.0)); // Does equality of near numbers.
angle.test_close(&Radians(2.0), 1.0);               // Does equality with test precision.

// For any other features, you can extract the decimal with:
angle.0;

println!("{}", angle);                              // Prints neatly "1.23r".
println!("{:?}", angle);                            // Prints in long form "1.2345".

Tuple Fields
0: Decimal
Implementations

impl Radians

pub fn abs(&self) -> Decimal

impl Radians

pub fn to_degrees(&self) -> Degrees

Converts to degrees.

pub fn to_hours(&self) -> Hours

Converts to hours format.

pub fn sin(&self) -> Decimal

Accessor for sin.

pub fn cos(&self) -> Decimal

Accessor for cos.