Struct star_tracker_lib::util::units::Vector3

pub struct Vector3 {
    pub x: Decimal,
    pub y: Decimal,
    pub z: Decimal,
}

A 3d cartesian vector/point.

This is to represent cartesian coordinates for doing angle calculations.
Cartesian is ideal over Equatorial in most circumstances as Equatorial has limited functionality.
Equatorial coordinates must first be converted to cartesian for most calculations.
Use Vector3 when you need to work with the coordinates.
Use Equatorial when you need to store it as it has 1 less value.

Equality

TestEqual and PartialEq are implemented to allow comparisons of equal vectors.
This ensures that if an equation and expected are off by a few bits, it is still valid.

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::aliases::DECIMAL_PRECISION;
use star_tracker_lib::util::test::TestEqual;
 
let vec_1 = Vector3{x: 1.0, y: 2.0, z: 3.0}; 
let vec_2 = Vector3{
	x: 1.0 + DECIMAL_PRECISION / 2.0, 
	y: 2.0 + DECIMAL_PRECISION / 2.0,
	z: 3.0 + DECIMAL_PRECISION / 2.0}; 
let vec_3 = Vector3{x: 2.0, y: 3.0, z: 4.0}; 

// A slight tolerance for standard equality.
assert_eq!(vec_1, vec_2);

// Can provide a tolerance if needed.
assert!( vec_1.test_close(&vec_3, 1.0001)); // True
assert!(!vec_1.test_close(&vec_3, 0.9999)); // False

Operators

There are standard operators for basic maths.

use star_tracker_lib::util::units::Vector3;
 
let vec = Vector3{x: 1.0, y: 2.0, z: 3.0}; 
// Addition / Subtraction 
assert_eq!(vec + vec, Vector3{x:  2.0, y:  4.0, z:  6.0});
assert_eq!(vec - vec, Vector3{x:  0.0, y:  0.0, z:  0.0});
assert_eq!(-vec,      Vector3{x: -1.0, y: -2.0, z: -3.0});

// Multiplication / Division
assert_eq!(vec * 2.0,   Vector3{x:  2.0, y:  4.0, z:  6.0});
assert_eq!(vec / 2.0,   Vector3{x:  0.5, y:  1.0, z: 1.5});

Other Operators

There are other operates provided for vector maths such as:

magnitude

normalize // mutates the actual object.

normalized // returns the result.

dot

cross

to_vector2

to_equatorial

to_matrix_row

to_matrix_column

to_matrix_column_homo // An extra row with 1 in it.

Fields
x: Decimaly: Decimalz: Decimal
Implementations

impl Vector3

pub fn magnitude(&self) -> Decimal

Finds the magnitude (length) of the vector.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::test::TestEqual;

let c = Vector3{x: 10.3, y: 23.1, z: 12.3};
assert!(c.magnitude().test_close(&28.124544440, 0.00001));

pub fn normalize(&mut self) -> Error<()>

Normalizes the vector so the magnitude/length is 1.
Returns Errors::NaN magnitude is 0.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::test::TestEqual;

let mut c = Vector3{x: 10.3, y: 23.1, z: 12.3};
let c_out = Vector3{x: 0.366228, y: 0.8213466, z: 0.43734};
c.normalize().expect("Will return Errors::NaN if magnitude is 0.");
assert!(c.test_close(&c_out, 0.0001));

pub fn normalized(&self) -> Error<Self>

Returns a new normalized vector of the current direction.
Returns Errors::NaN magnitude is 0.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::test::TestEqual;

let mut c = Vector3{x: 10.3, y: 23.1, z: 12.3};
let c_out = Vector3{x: 0.366228, y: 0.8213466, z: 0.43734};
assert!(c.normalized().expect("NaN if magnitude is 0.").test_close(&c_out, 0.0001));

pub fn cross(&self, other: Vector3) -> Vector3

Finds the cross product between self and the input object.

Example

use star_tracker_lib::util::units::Vector3;

let a = Vector3 { x: -1.0, y: 2.0, z: 10.0 };
let b = Vector3 { x: 9.0, y: 3.0, z: -4.0 };
assert_eq!(a.cross(b), Vector3{x: -38.0, y: 86.0, z: -21.0});
assert_eq!(b.cross(a), Vector3{x: 38.0, y: -86.0, z: 21.0});

pub fn dot(&self, other: Vector3) -> Decimal

Finds the dot product between the Vector3 points.

Example

use star_tracker_lib::util::units::Vector3;

let a = Vector3 { x: 2.0, y: 3.0, z: 4.0 };
let b = Vector3 { x: 5.0, y: 6.0, z: 7.0 };
assert_eq!(a.dot(b), 56.0);

pub fn angle_distance(&self, oth: Vector3) -> Radians

Finds the angle between the 2 points on a sphere.
The great circle distance.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Degrees;

let mut car1 = Vector3 { x: 0.0, y: 1.0, z: 0.0 };
let mut car2 = Vector3 { x: 1.0, y: 0.0, z: 0.0 };
assert_eq!(car1.angle_distance(car2), Degrees(90.0).to_radians());

pub fn to_matrix_column(&self) -> Matrix<3, 1>

Converts the coordinates into a column matrix form (vertical set of rows).

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let c = Vector3 {x: 1.0, y: 2.0, z: 3.0};
let m: Matrix<3,1> = c.to_matrix_column();
assert_eq!(c.x, m.get(MatPos{row: 0, col: 0}));
assert_eq!(c.y, m.get(MatPos{row: 1, col: 0}));
assert_eq!(c.z, m.get(MatPos{row: 2, col: 0}));

pub fn to_matrix_row(&self) -> Matrix<1, 3>

Converts the coordinates into a row matrix form (horizontal set of columns).

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let c = Vector3 {x: 1.0, y: 2.0, z: 3.0};
let m: Matrix<1,3> = c.to_matrix_row();
assert_eq!(c.x, m.get(MatPos{row: 0, col: 0}));
assert_eq!(c.y, m.get(MatPos{row: 0, col: 1}));
assert_eq!(c.z, m.get(MatPos{row: 0, col: 2}));

pub fn to_matrix_column_homo(&self) -> Matrix<4, 1>

Converts the coordinates into a column matrix form (vertical set of rows). This is in homogeneous form for matrix transformations.

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Matrix;
use star_tracker_lib::util::units::MatPos;

let c = Vector3 {x: 1.0, y: 2.0, z: 3.0};
let m: Matrix<4,1> = c.to_matrix_column_homo();
assert_eq!(c.x, m.get(MatPos{row: 0, col: 0}));
assert_eq!(c.y, m.get(MatPos{row: 1, col: 0}));
assert_eq!(c.z, m.get(MatPos{row: 2, col: 0}));
assert_eq!(1.0, m.get(MatPos{row: 3, col: 0}));

pub fn to_vector2(&self) -> Vector2

Converts self to vector2 (LOSSY).

Example

use star_tracker_lib::util::units::Vector3;
use star_tracker_lib::util::units::Vector2;

let vec_3 = Vector3{x: 12.3, y: 23.4, z: 34.5};
let vec_2 = vec_3.to_vector2();
assert_eq!(vec_2.x, vec_3.x);
assert_eq!(vec_2.y, vec_3.y);

pub fn to_equatorial(&self) -> Equatorial

Converts Vector3 to Equatorial coordinates.

Example

use star_tracker_lib::util::aliases::M_PI;
use star_tracker_lib::util::units::{Equatorial, Vector3, Radians};
use star_tracker_lib::util::aliases::Decimal;
use star_tracker_lib::util::test::TestEqual;

let mut c = Vector3 { x: 0.5, y: 0.5, z: -0.7071067812 };
let mut e = c.to_equatorial();
let mut compare = Equatorial{ra: Radians(M_PI / 4.0), dec: Radians(-M_PI / 4.0)};
assert!(compare.test_close(&e, 0.0000001));

c = Vector3 { x: 3.0, y: 4.0, z: 5.0 };
e = c.to_equatorial();
compare = Equatorial{ra: Radians(0.927295218), dec: Radians(0.78539816)};
assert!(compare.test_close(&e, 0.0000001));